% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/add_ancestor.R, R/add_death.R,
%   R/add_disease.R, R/add_disease_genes.R, R/add_evidence.R,
%   R/add_gene_frequency.R, R/add_genes.R, R/add_hpo_definition.R,
%   R/add_hpo_id.R, R/add_hpo_name.R, R/add_info_content.R, R/add_mondo.R,
%   R/add_ndisease.R, R/add_omop.R, R/add_onset.R, R/add_ont_lvl.R,
%   R/add_pheno_frequency.R, R/add_prevalence.R, R/add_severity.R, R/add_tier.R
\name{add_}
\alias{add_}
\alias{add_ancestor}
\alias{add_death}
\alias{add_disease}
\alias{add_disease_genes}
\alias{add_evidence}
\alias{add_gene_frequency}
\alias{add_genes}
\alias{add_hpo_definition}
\alias{add_hpo_id}
\alias{add_hpo_name}
\alias{add_info_content}
\alias{add_mondo}
\alias{add_ndisease}
\alias{add_omop}
\alias{add_onset}
\alias{add_ont_lvl}
\alias{add_pheno_frequency}
\alias{add_prevalence}
\alias{add_severity}
\alias{add_tier}
\title{Add functions}
\usage{
add_ancestor(
  phenos,
  lvl = 2,
  hpo = get_hpo(),
  keep_descendants = NULL,
  remove_descendants = NULL,
  force_new = FALSE
)

add_death(
  phenos,
  keep_deaths = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  agg_by = NULL
)

add_disease(
  phenos,
  extra_cols = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  add_definitions = FALSE
)

add_disease_genes(phenos, all.x = TRUE, verbose = TRUE)

add_evidence(
  phenos,
  keep_evidence = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  agg_by = c("disease_id", "gene_symbol"),
  default_score = 1
)

add_gene_frequency(
  phenotype_to_genes = load_phenotype_to_genes(),
  gene_frequency_threshold = NULL,
  all.x = TRUE,
  verbose = TRUE
)

add_genes(
  phenos = NULL,
  phenotype_to_genes = load_phenotype_to_genes(),
  hpo = get_hpo(),
  by = c("hpo_id", "disease_id"),
  gene_col = "gene_symbol",
  all.x = FALSE,
  allow.cartesian = FALSE
)

add_hpo_definition(
  phenos,
  line_length = FALSE,
  use_api = FALSE,
  verbose = TRUE
)

add_hpo_id(phenos, hpo = get_hpo(), phenotype_to_genes = NULL)

add_hpo_name(phenos, hpo = get_hpo(), phenotype_to_genes = NULL)

add_info_content(phenos, hpo = get_hpo())

add_mondo(phenos, input_col = "disease_id", map_to = "hpo", ...)

add_ndisease(
  phenos,
  pheno_ndiseases_threshold = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  verbose = TRUE
)

add_omop(
  phenos,
  input_col = "hpo_id",
  all.x = TRUE,
  allow.cartesian = FALSE,
  force_new = FALSE,
  verbose = TRUE
)

add_onset(
  phenos,
  keep_onsets = NULL,
  agg_by = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE
)

add_ont_lvl(
  phenos,
  hpo = get_hpo(),
  absolute = TRUE,
  keep_ont_levels = NULL,
  ...
)

add_pheno_frequency(
  phenos,
  pheno_frequency_threshold = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE
)

add_prevalence(
  phenos,
  input_col = "disease_id",
  drop_na = TRUE,
  method = "orphanet"
)

add_severity(
  phenos,
  hpo = get_hpo(),
  all.x = TRUE,
  allow.cartesian = FALSE,
  severity_threshold = NULL
)

add_tier(
  phenos,
  all.x = TRUE,
  include_disease_characteristics = TRUE,
  auto_assign = TRUE,
  hpo = get_hpo(),
  keep_tiers = NULL,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{A data.table containing HPO IDs and other metadata.}

\item{lvl}{How many levels deep into the ontology to get ancestors from.
For example:
\itemize{
\item{1: }{"All"}
\item{2: }{"Phenotypic abnormality"}
\item{3: }{"Abnormality of the nervous system"}
\item{4: }{"Abnormality of nervous system physiology"}
\item{5: }{"Neurodevelopmental abnormality" or "Behavioral abnormality"}
}}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{remove_descendants}{Terms whose descendants should be removed 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{force_new}{Force a new query to the OARD API instead of
using pre-downloaded data.}

\item{keep_deaths}{The age of death associated with each HPO ID to keep.
If >1 age of death is associated with the term,
only the earliest age is considered.
See \link{add_death} for details.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{agg_by}{Column to aggregate age of onset metadata by.}

\item{extra_cols}{Extra metadata columns from the"phenotype.hpoa"
annotations file to include.
See
\href{https://hpo-annotation-qc.readthedocs.io/en/latest/annotationFormat.html}{
here for column descriptions}.}

\item{add_definitions}{Add disease definitions using \link{add_mondo}.}

\item{verbose}{Print messages.}

\item{keep_evidence}{The evidence scores of each gene-disease
association to keep.}

\item{default_score}{Default evidence score to
apply to gene-disease associations that are present in the HPO annotations
but don't have evidence scores in the GenCC annotations.}

\item{phenotype_to_genes}{Output of
\link{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{gene_frequency_threshold}{Only keep genes with frequency
above the set threshold. Frequency ranges from 0-100 where 100 is
a gene that occurs 100\% of the time in a given phenotype.
Include \code{NA} if you wish to retain genes that
do not have any frequency data.
See \link{add_gene_frequency} for details.}

\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{gene_col}{Name of the gene column.}

\item{line_length}{The number of desired words per line \<int\>}

\item{use_api}{Get definitions from the HPO API,
as opposed to a static local dataset.}

\item{input_col}{Name of the column containing the disease or phenotype IDs.}

\item{map_to}{Mapping outputs to include
(from Mondo IDs to another database's IDs).}

\item{...}{
  Arguments passed on to \code{\link[KGExplorer:map_]{KGExplorer::map_mondo}}, \code{\link[KGExplorer:get_]{KGExplorer::get_ontology_levels}}
  \describe{
    \item{\code{dat}}{\link[data.table]{data.table} with genes.}
    \item{\code{output_col}}{Column name of output IDs.}
    \item{\code{to}}{Character vector of database(s) to map IDs to.
When not \code{"mondo"}, can supply multiple alternative databases to map to
 (e.g. \code{c("OMIM","Orphanet","DECIPHER")}).}
    \item{\code{map_types}}{Mapping types to include.}
    \item{\code{top_n}}{Top number of mappings to return per \code{top_by} grouping.
Set to \code{NULL} to skip this step.}
    \item{\code{add_name}}{Logical, if TRUE, add mondo name column.}
    \item{\code{save_dir}}{Directory to save cached data.}
    \item{\code{ont}}{An ontology of class \link[simona]{ontology_DAG}.}
    \item{\code{terms}}{Term IDs to include. Can alternatively be an integer, 
which will be used to randomly sample N terms from the data.}
    \item{\code{remove_terms}}{Character vector of term IDs to exclude.}
    \item{\code{reverse}}{If \code{TRUE}, ontology
level numbers with be revered such that the level of the parent terms
are larger than the child terms.}
  }}

\item{pheno_ndiseases_threshold}{Filter phenotypes by the
maximum number of diseases they are associated with.}

\item{keep_onsets}{The age of onset associated with each HPO ID to keep.
If >1 age of onset is associated with the term,
only the earliest age is considered.
See \link[HPOExplorer]{add_onset} for details.}

\item{absolute}{Make the levels absolute in the sense that they consider
the entire ontology (\code{TRUE}).
Otherwise, levels will be relative to only the terms that are in
 the provided subset of \code{terms} AND are directly adjacent (connected)
 to a given cluster of terms (\code{FALSE}).}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link{add_ont_lvl} for details.}

\item{pheno_frequency_threshold}{Only keep phenotypes with frequency
above the set threshold. Frequency ranges from 0-100 where 100 is
a phenotype that occurs 100\% of the time in all associated diseases.
Include \code{NA} if you wish to retain phenotypes that
do not have any frequency data.
See \link[HPOExplorer]{add_pheno_frequency} for details.}

\item{drop_na}{Whether to drop rows with missing prevalence data.}

\item{method}{One of "orphanet" or "oard".}

\item{severity_threshold}{Only keep phenotypes with a mean
severity score (averaged across multiple associated diseases) below the
set threshold. The severity score ranges from 1-4 where 1 is the MOST severe.
Include \code{NA} if you wish to retain phenotypes that
do not have any severity score.}

\item{include_disease_characteristics}{Include \code{phenotypes} that
are also high-level \code{include_disease_characteristics}.}

\item{auto_assign}{Automatically assing HPO IDs to Tiers by conducting
regex searches for keywords that appear in the term name,
or the names of its descendants or ancestors.}

\item{keep_tiers}{Tiers from \link[HPOExplorer]{hpo_tiers} to keep.
Include \code{NA} if you wish to retain phenotypes that
do not have any Tier assignment.}
}
\value{
Annotated data.

phenos data.table with extra columns:
\itemize{
\item{"AgeOfDeath": }{AgeOfDeath HPO IDs of disease phenotypes associated
with the target hpo_id phenotype.}
\item{"AgeOfDeath_names": }{AgeOfDeath HPO names of disease phenotypes
associated with the target hpo_id phenotype.}
\item{"AgeOfDeath_counts": }{The number of times each term in
"AgeOfDeath_names" appears across associated disease phenotypes.}
\item{"AgeOfDeath_score_mean": }{Mean age of death score.}
\item{"AgeOfDeath_score_min": }{Minimum age of death score.}
\item{"AgeOfDeath_top": }{The most common age of death term.}
\item{"AgeOfDeath_earliest": }{The earliest age of death.}
\item{"AgeOfDeath_latest": }{The latest age of death.}
}

phenos data.table with extra columns:
\itemize{
\item{"evidence_score_min": }{Minimum evidence score.}
\item{"evidence_score_max": }{Maximum evidence score.}
\item{"evidence_score_mean": }{Mean evidence score.}
}

phenos data.table with extra column

A named vector of HPO term descriptions.

phenos data.table with extra column

phenos data.table with extra columns.

phenos data.table with extra columns

phenos data.table with extra columns:
\itemize{
\item{"onset": }{onset HPO IDs of disease phenotypes associated
with the target hpo_id phenotype.}
\item{"onset_names": }{onset HPO names of disease phenotypes associated
with the target hpo_id phenotype.}
\item{"onset_counts": }{The number of times each term in
"onset_names" appears across associated disease phenotypes.}
\item{"onset_score_mean": }{Mean onset score.}
\item{"onset_score_min": }{Minimum onset score.}
\item{"onset_top": }{The most common onset term.}
\item{"onset_earliest": }{The earliest age of onset.}
\item{"onset_latest": }{The latest age of onset.}
}

phenos data.table with extra column

phenos data.table with extra column

phenos data.table with extra columns

phenos data.table with extra column
}
\description{
Functions to add metadata to data.table objects.
}
\section{Functions}{
\itemize{
\item \code{add_ancestor()}: add_
Add ancestor

Assign each HPO ID to the higher-order ancestral term that it is part of.

\item \code{add_death()}: add_
Add age of death

Add age of death for each HPO ID.
AgeOfDeath IDs and assigned "AgeOfDeath_score" values:
\itemize{
\item{HP:0005268 }{"Miscarriage" (AgeOfDeath_score=1)}
\item{HP:0003826 }{"Stillbirth" (AgeOfDeath_score=1)}
\item{HP:0034241 }{"Prenatal death" (AgeOfDeath_score=1)}
\item{HP:0003811 }{"Neonatal death" (AgeOfDeath_score=2)}
\item{HP:0001522 }{"Death in infancy" (AgeOfDeath_score=3)}
\item{HP:0003819 }{"Death in childhood" (AgeOfDeath_score=4)}
\item{HP:0011421 }{"Death in adolescence" (AgeOfDeath_score=5)}
\item{HP:0100613 }{"Death in early adulthood" (AgeOfDeath_score=6)}
\item{HP:0033764 }{"Death in middle age" (AgeOfDeath_score=7)}
\item{HP:0033763 }{"Death in adulthood" (AgeOfDeath_score=7)}
\item{HP:0033765 }{"Death in late adulthood" (AgeOfDeath_score=8)}
}

\item \code{add_disease()}: add_
Add diseases

Annotate each HPO term with diseases that they are associated with.

\item \code{add_disease_genes()}: add_
Add disease genes

Add genes that overlap between an HPO ID and an associated phenotype.

\item \code{add_evidence()}: add_
Add evidence

Add the strength of evidence supporting each gene-disease association.
\href{https://thegencc.org/faq.html#validity-termsdelphi-survey}{
Delphi survey} evidence classification IDs and
assigned "evidence_score" values:
\itemize{
\item{GENCC:100001 }{"Definitive" (evidence_score=6)}
\item{GENCC:100002 }{"Strong" (evidence_score=5)}
\item{GENCC:100003 }{"Moderate" (evidence_score=4)}
\item{GENCC:100009 }{"Supportive" (evidence_score=3)}
\item{GENCC:100004 }{"Limited" (evidence_score=2)}
\item{GENCC:100005 }{"Disputed Evidence" (evidence_score=1)}
\item{GENCC:100006 }{"Refuted Evidence" (evidence_score=0)}
\item{GENCC:100008 }{"No Known Disease Relationship" (evidence_score=0)}
}

\item \code{add_gene_frequency()}: add_
Add gene frequency

Add gene-level frequency, i.e. how often mutations in a given gene
are associated with a given phenotype.
Numeric frequency columns are on a 0-100\% scale.

\item \code{add_genes()}: add_
Add genes

Add genes associated with each phenotype
(in the context of a particular disease).

\item \code{add_hpo_definition()}: add_
Get term definition

This function accesses the HPO API to get a description/definition of an
HPO term. If a \code{line_length} \> 0 is passed to the function, it will add
newlines every nth word. This can be useful when displaying the description
in plots with limited space.

\item \code{add_hpo_id()}: add_
Add HPO ID column to dataframe

Adds the HPO term ID column "hpo_id".

\item \code{add_hpo_name()}: add_
Add HPO name column to dataframe

Adds the HPO term name column "hpo_name".

\item \code{add_info_content()}: add_
Add information content

Add a column containing the information content score  for each HPO ID.

\item \code{add_mondo()}: add_
Add Mondo metadata

Add Mondo metadata (MONDO ID mappings, names, and definitions) for diseases
using files from their respective databases:
e.g. OMIM, DECIPHER, Orphanet.

\item \code{add_ndisease()}: add_
Add N diseases

Annotate each HPO term with the total number of disease
they are associated with.

\item \code{add_omop()}: add_
Add OMOP

Add metadata from \href{https://mondo.monarchinitiative.org/}{MONDO},
including:
\itemize{
\item{mondo_id: }{MONDO term ID.}
\item{mondo_name: }{MONDO term name.}
\item{mondo_def: }{MONDO term definition.}
}

\item \code{add_onset()}: add_
Add age of onset

Add age of onset for each HPO ID.
onset IDs and assigned "onset_score" values:
\itemize{
\item{HP:0011461 }{"Fetal onset" (onset_score=1)}
\item{HP:0030674 }{"Antenatal onset" (onset_score=2)}
\item{HP:0003577 }{"Congenital onset" (onset_score=3)}
\item{HP:0003623 }{"Neonatal onset" (onset_score=4)}
\item{HP:0003593 }{"Infantile onset" (onset_score=5)}
\item{HP:0011463 }{"Childhood onset" (onset_score=6)}
\item{HP:0003621 }{"Juvenile onset" (onset_score=7)}
\item{HP:0011462 }{"Young adult onset" (onset_score=8)}
\item{HP:0003581 }{"Adult onset" (onset_score=9)}
\item{HP:0003596 }{"Middle age onset" (onset_score=10)}
\item{HP:0003584 }{"Late onset" (onset_score=11)}
}

\item \code{add_ont_lvl()}: add_
Add ontology level

Add the relative ontology level for each HPO ID.

\item \code{add_pheno_frequency()}: add_
Add phenotype frequency

Add phenotype-level frequency, i.e.
how often a phenotype occurs in a given disease.

\item \code{add_prevalence()}: add_
Add prevalence

Add a column containing the prevalence score for each disease ("disease_id")
or phenotype ("hpo_id").

\item \code{add_severity()}: add_
Add HPO modifiers

Annotate each HPO with modifier terms, including
(but not limited to) progression and severity ratings.
In order of increasing severity:
\itemize{
\item{HP:0012825 }{"Mild" (Severity_score=4)}
\item{HP:0012827 }{"Borderline" (Severity_score=3)}
\item{HP:0012828 }{"Severe" (Severity_score=2)}
\item{HP:0012829}{"Profound" (Severity_score=1)}
}

\item \code{add_tier()}: add_
Add severity Tiers

Add severity Tier for each HPO ID, in accordance with the
rating system provided by
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4262393/}{
Lazarin et al (2014)}.
In order of increasing severity:
\itemize{
\item{Tier 4 }{Reduced fertility}
\item{Tier 3 }{Sensory impairment: vision,
Immunodeficiency/cancer,
Sensory impairment: hearing,
Sensory impairment: touch, other (including pain),
Mental illness,
Dysmorphic features}
\item{Tier 2 }{Shortened life span: premature adulthood,
Impaired mobility,
Internal physical malformation}
\item{Tier 1 }{Shortened life span: infancy,
Shortened life span: childhood/adolescence,
Intellectual disability}
}

}}
\examples{
phenos <- example_phenos()
phenos2 <- add_ancestor(phenos = phenos, lvl=5)
phenos <- example_phenos()
phenos2 <- add_death(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_disease(phenos = phenos)
\dontrun{
phenos <- load_phenotype_to_genes()
phenos2 <- add_severity(phenos = phenos)
}
phenos <- load_phenotype_to_genes()
phenos2 <- add_evidence(phenos = phenos)
phenotype_to_genes <- load_phenotype_to_genes()[seq(1000),]
phenos2 <- add_gene_frequency(phenotype_to_genes = phenotype_to_genes)
phenos <- example_phenos()
phenos2 <- add_genes(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_hpo_definition(phenos = phenos)
phenotype_to_genes <- load_phenotype_to_genes()
phenos <- unique(phenotype_to_genes[,c("hpo_id","hpo_name")])
phenos2 <- add_hpo_id(phenos=phenos)
phenos <- example_phenos()
phenos2 <- add_hpo_name(phenos=phenos)
phenos <- example_phenos()
phenos2 <- add_info_content(phenos = phenos)
phenos <- load_phenotype_to_genes(3)[seq(1000)]
phenos2 <- add_mondo(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_ndisease(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_omop(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_onset(phenos = phenos)
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_ont_lvl(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_pheno_frequency(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_prevalence(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_severity(phenos = phenos)
phenos <- example_phenos()
phenos2 <- add_tier(phenos = phenos)
}
\concept{add_}
