% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ancestor.R
\name{add_ancestor}
\alias{add_ancestor}
\title{Add ancestor}
\usage{
add_ancestor(
  phenos,
  lvl = 3,
  hpo = get_hpo(),
  remove_descendants = NULL,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{lvl}{How many levels deep into the ontology to get ancestors from.
For example:
\itemize{
\item{1: }{"All"}
\item{2: }{"Phenotypic abnormality"}
\item{3: }{"Abnormality of the nervous system"}
\item{4: }{"Abnormality of nervous system physiology"}
\item{5: }{"Neurodevelopmental abnormality" or "Behavioral abnormality"}
}}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{remove_descendants}{Remove HPO terms that are descendants of a given
ancestral HPO term. Ancestral terms be provided as a character vector of
phenotype names (e.g. \code{c("Clinical course")}),
HPO IDs (e.g. \code{"HP:0031797" }) or a mixture of the two.
See \link[HPOExplorer]{add_ancestor} for details.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Assign each HPO ID to the higher-order ancestral term that it is part of.
}
\examples{
phenos <- example_phenos()
phenos2 <- add_ancestor(phenos = phenos, lvl=5)
}
