% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_death.R
\name{add_death}
\alias{add_death}
\title{Add age of death}
\usage{
add_death(
  phenos,
  keep_deaths = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  agg_by = NULL,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{keep_deaths}{The age of death associated with each HPO ID to keep.
If >1 age of death is associated with the term,
only the earliest age is considered.
See \link[HPOExplorer]{add_death} for details.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{agg_by}{Column to aggregate age of death metadata by.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra columns:
\itemize{
\item{"AgeOfDeath": }{AgeOfDeath HPO IDs of disease phenotypes associated
with the target HPO_ID phenotype.}
\item{"AgeOfDeath_names": }{AgeOfDeath HPO names of disease phenotypes
associated with the target HPO_ID phenotype.}
\item{"AgeOfDeath_counts": }{The number of times each term in
"AgeOfDeath_names" appears across associated disease phenotypes.}
\item{"AgeOfDeath_score_mean": }{Mean age of death score.}
\item{"AgeOfDeath_score_min": }{Minimum age of death score.}
\item{"AgeOfDeath_top": }{The most common age of death term.}
\item{"AgeOfDeath_earliest": }{The earliest age of death.}
\item{"AgeOfDeath_latest": }{The latest age of death.}
}
}
\description{
Add age of death for each HPO ID.
AgeOfDeath IDs and assigned "AgeOfDeath_score" values:
\itemize{
\item{HP:0005268 }{"Miscarriage" (AgeOfDeath_score=1)}
\item{HP:0003826 }{"Stillbirth" (AgeOfDeath_score=1)}
\item{HP:0034241 }{"Prenatal death" (AgeOfDeath_score=1)}
\item{HP:0003811 }{"Neonatal death" (AgeOfDeath_score=2)}
\item{HP:0001522 }{"Death in infancy" (AgeOfDeath_score=3)}
\item{HP:0003819 }{"Death in childhood" (AgeOfDeath_score=4)}
\item{HP:0011421 }{"Death in adolescence" (AgeOfDeath_score=5)}
\item{HP:0100613 }{"Death in early adulthood" (AgeOfDeath_score=6)}
\item{HP:0033764 }{"Death in middle age" (AgeOfDeath_score=7)}
\item{HP:0033763 }{"Death in adulthood" (AgeOfDeath_score=7)}
\item{HP:0033765 }{"Death in late adulthood" (AgeOfDeath_score=8)}
}
}
\examples{
phenos <- example_phenos()
phenos2 <- add_death(phenos = phenos)
}
