% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_disease.R
\name{add_disease}
\alias{add_disease}
\title{Add diseases}
\usage{
add_disease(
  phenos,
  extra_cols = NULL,
  include_mondo = TRUE,
  all.x = TRUE,
  allow.cartesian = FALSE,
  add_definitions = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{extra_cols}{Extra metadata columns from the"phenotype.hpoa"
annotations file to include.
See
\href{https://hpo-annotation-qc.readthedocs.io/en/latest/annotationFormat.html}{
here for column descriptions}.}

\item{include_mondo}{Add IDs/names/definitions from MONDO via
\link[HPOExplorer]{add_mondo}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{add_definitions}{Add disease definitions using
\link[HPOExplorer]{add_disease_definition}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra columns
}
\description{
Annotate each HPO term with diseases that they are associated with.
}
\examples{
phenos <- example_phenos()
phenos2 <- add_disease(phenos = phenos)
}
