% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_disease_genes.R
\name{add_disease_genes}
\alias{add_disease_genes}
\title{Add disease genes}
\usage{
add_disease_genes(phenos, all.x = TRUE, verbose = TRUE)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra columns
}
\description{
Add genes that overlap between an HPO ID and an associated phenotype.
}
\examples{
phenos <- load_phenotype_to_genes()[,HPO_ID:=ID]
phenos2 <- add_modifier(phenos = phenos)
}
