% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_genes.R
\name{add_genes}
\alias{add_genes}
\title{Add genes}
\usage{
add_genes(
  phenos = NULL,
  phenotype_to_genes = load_phenotype_to_genes(),
  hpo = get_hpo(),
  by = c("HPO_ID", "DatabaseID"),
  gene_col = "Gene",
  all.x = FALSE,
  allow.cartesion = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{verbose}{Print messages.}
}
\description{
Add genes associated with each phenotype
(in the context of a particular disease).
}
\examples{
phenos <- example_phenos()
phenos2 <- add_genes(phenos = phenos)
}
