% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hpo_definition.R
\name{add_hpo_definition}
\alias{add_hpo_definition}
\title{Get term definition}
\usage{
add_hpo_definition(
  phenos,
  line_length = FALSE,
  use_api = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{line_length}{The number of desired words per line \<int\>}

\item{use_api}{Get definitions from the HPO API,
as opposed to a static local dataset.}

\item{verbose}{Print messages.}
}
\value{
A named vector of HPO term descriptions.
}
\description{
This function accesses the HPO API to get a description/definition of an
HPO term. If a \code{line_length} \> 0 is passed to the function, it will add
newlines every nth word. This can be useful when displaying the description
in plots with limited space.
}
\examples{
phenos <- data.table::data.table(HPO_ID=get_hpo()$id[seq_len(10)])
phenos2 <- add_hpo_definition(phenos = phenos)
}
