% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hpo_id.R
\name{add_hpo_id}
\alias{add_hpo_id}
\title{Add HPO ID column to dataframe}
\usage{
add_hpo_id(phenos, hpo = get_hpo(), phenotype_to_genes = NULL, verbose = FALSE)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{verbose}{Print messages.}
}
\value{
The subset of EWCE results dataframe with a HPO Id column added.
}
\description{
This adds the HPO term id column to the subset of ewce results data
to be plotted
in the cell select app. It also checks if it is a valid HPO term id to
prevent error and adds
a boolean column where TRUE if term is valid. If the HPO Id is not correct,
it caused an error in the ontologyPlot package.
}
\examples{
phenotype_to_genes <- load_phenotype_to_genes()
phenos <- unique(phenotype_to_genes[,c("hpo_id","hpo_name")])
phenos2 <- add_hpo_id(phenos=phenos)
}
