% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hpo_synonym.R
\name{add_hpo_synonym}
\alias{add_hpo_synonym}
\title{Add HPO ID synonyms}
\usage{
add_hpo_synonym(
  phenos,
  to = "UMLS",
  to_col = paste0(to, "_ID"),
  hpo = get_hpo(),
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{to}{The ontology to map to. Default is \code{UMLS}.}

\item{to_col}{The name of the new column to add the synonyms to.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{verbose}{Print messages.}
}
\value{
A data.table with the HPO ID synonyms in a new column with
the same name as \code{to}.
}
\description{
Map HPO IDs to synonynmous IDs from other ontologies
(e.g. UMLS, SNOMEDCT_US, MSH, ICD-10)
}
\examples{
phenos <- HPOExplorer::example_phenos()
phenos2 <- add_hpo_synonym(phenos)
}
