% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_info_content.R
\name{add_info_content}
\alias{add_info_content}
\title{Add information content}
\usage{
add_info_content(phenos, hpo = get_hpo(), verbose = TRUE)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add a column containing the information content score  for each HPO ID.
}
\examples{
phenos <- example_phenos()
phenos2 <- add_info_content(phenos = phenos)
}
