% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ndisease.R
\name{add_ndisease}
\alias{add_ndisease}
\title{Add N diseases}
\usage{
add_ndisease(
  phenos,
  pheno_ndiseases_threshold = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{pheno_ndiseases_threshold}{Filter phenotypes by the
maximum number of diseases they are associated with.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra columns
}
\description{
Annotate each HPO term with the total number of disease
they are associated with.
}
\examples{
phenos <- example_phenos()
phenos2 <- add_ndisease(phenos = phenos)
}
