% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ont_lvl.R
\name{add_ont_lvl}
\alias{add_ont_lvl}
\title{Add ontology level}
\usage{
add_ont_lvl(
  phenos,
  hpo = get_hpo(),
  adjacency = adjacency_matrix(terms = phenos$HPO_ID, hpo = hpo),
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{AN adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add the relative ontology level for each HPO ID.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_ont_lvl(phenos = phenos)
}
