% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pheno_frequency.R
\name{add_pheno_frequency}
\alias{add_pheno_frequency}
\title{Add phenotype frequency}
\usage{
add_pheno_frequency(phenos, all.x = TRUE, verbose = TRUE)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add phenotype-level frequency, i.e.
how often a phenotype occurs in a given disease.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_pheno_frequency(phenos = phenos)
}
