% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prevalence.R
\name{add_prevalence}
\alias{add_prevalence}
\title{Add prevalence}
\usage{
add_prevalence(
  phenos,
  id_col = "disease_id",
  drop_na = TRUE,
  verbose = TRUE,
  method = "orphanet"
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{id_col}{Name of the column containing the disease or phenotype IDs.}

\item{drop_na}{Whether to drop rows with missing prevalence data.}

\item{verbose}{Print messages.}

\item{method}{One of "orphanet" or "oard".}
}
\value{
phenos data.table with extra column
}
\description{
Add a column containing the prevalence score for each disease ("disease_id")
or phenotype ("hpo_id").
}
\examples{
phenos <- example_phenos()
phenos2 <- add_prevalence(phenos = phenos)
}
