% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_severity.R
\name{add_severity}
\alias{add_severity}
\title{Add HPO modifiers}
\usage{
add_severity(
  phenos,
  hpo = get_hpo(),
  all.x = TRUE,
  severity_threshold = NULL,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{severity_threshold}{Only keep phenotypes with a mean
severity score (averaged across multiple associated diseases) below the
set threshold. The severity score ranges from 1-4 where 1 is the MOST severe.
Include \code{NA} if you wish to retain phenotypes that
do not have any severity score.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra columns
}
\description{
Annotate each HPO with modifier terms, including
(but not limited to) progression and severity ratings.
In order of increasing severity:
\itemize{
\item{HP:0012825 }{"Mild" (Severity_score=4)}
\item{HP:0012827 }{"Borderline" (Severity_score=3)}
\item{HP:0012828 }{"Severe" (Severity_score=2)}
\item{HP:0012829}{"Profound" (Severity_score=1)}
}
}
\examples{
phenos <- example_phenos()
phenos2 <- add_severity(phenos = phenos)
}
