% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tier.R
\name{add_tier}
\alias{add_tier}
\title{Add severity Tiers}
\usage{
add_tier(
  phenos,
  all.x = TRUE,
  include_disease_characteristics = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{include_disease_characteristics}{Include \code{phenotypes} that
are also high-level \code{include_disease_characteristics}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add age of onset for each HPO ID.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_tier(phenos = phenos)
}
