% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Adjacency matrix}
\usage{
adjacency_matrix(
  hpo = get_hpo(),
  terms = unique(hpo$id),
  method = "HPOExplorer",
  verbose = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{terms}{Character vector of HPO IDs.}

\item{method}{Method to use to create the adjacency matrix.}

\item{verbose}{Print messages.}
}
\value{
adjacency matrix
}
\description{
Create adjacency matrix of HPO child-parent relationships.
}
\examples{
terms <- get_hpo()$id[seq(1000)]
adjacency <- adjacency_matrix(terms = terms)
}
