% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_phenos.R
\name{annotate_phenos}
\alias{annotate_phenos}
\title{Annotate phenotypes}
\usage{
annotate_phenos(
  phenos,
  hpo = get_hpo(),
  adjacency = NULL,
  add_ont_lvl_absolute = TRUE,
  add_ont_lvl_relative = FALSE,
  add_info_contents = FALSE,
  add_description = TRUE,
  add_disease_data = FALSE,
  add_ndiseases = add_disease_data,
  add_onsets = add_disease_data,
  add_deaths = add_disease_data,
  add_pheno_frequencies = add_disease_data,
  add_tiers = add_disease_data,
  add_severities = add_disease_data,
  add_hoverboxes = FALSE,
  columns = list_columns(),
  interactive = TRUE,
  verbose = TRUE
)
}
\description{
Annotate phenotypes \link[data.table]{data.table} without various types
of metadata.
}
\examples{
phenos <- example_phenos()
phenos2 <- annotate_phenos(phenos)
}
