% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_annot_gpt.R
\name{check_annot_gpt}
\alias{check_annot_gpt}
\title{Check annotations from GPT}
\usage{
check_annot_gpt(path)
}
\arguments{
\item{path}{Path to annotations file.}
}
\value{
Named list
}
\description{
Check chatGPT phenotype annotations using a several metrics.
}
\keyword{internal}
