% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_map}
\alias{disease_map}
\title{Human Phenotype Ontology: Disease ID to Disease Name and Disease Description}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8631 rows and 3 columns.
}
\source{
\code{
 dat <- load_phenotype_to_genes()
 out <- map_disease(dat)
 disease_map <- out[,list(disease_id,disease_name,disease_description)]|>unique()
 usethis::use_data(disease_map, overwrite = TRUE)
 }
@format data.table
@usage data("disease_map")
}
\usage{
disease_map
}
\description{
Mapping of HPO disease ID (disease_id) to Disease Name and
Disease Description.
}
\keyword{datasets}
