% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_phenos.R
\name{example_phenos}
\alias{example_phenos}
\title{Example phenotypes dataframe}
\usage{
example_phenos(i = seq(10))
}
\arguments{
\item{i}{Indices of HPO IDs to use.}
}
\value{
phenotype data.table
}
\description{
Create a minimal example of a phenos dataframe.
}
\examples{
phenos <- example_phenos()
}
