% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/filter_descendants.R
\name{filter_}
\alias{filter_}
\alias{filter_descendants}
\title{Filter functions}
\usage{
filter_descendants(
  phenos,
  keep_descendants = NULL,
  remove_descendants = NULL,
  hpo = get_hpo()
)
}
\arguments{
\item{phenos}{A data.table containing HPO IDs and other metadata.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{remove_descendants}{Terms whose descendants should be removed 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}
}
\value{
Filtered data.

data.table of phenotypes, with additional columns:
"ancestor", "ancestor_id"
}
\description{
Functions to filter data.table objects.
}
\section{Functions}{
\itemize{
\item \code{filter_descendants()}: filter_
Filter descendants

Subset a \code{phenos} data.table to only
descendants of an ancestor HPO ID term.

}}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- filter_descendants(phenos = phenos,
                              keep_descendants = "Motor delay")
}
\concept{filter_}
