% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_hpo_ids.R
\name{fix_hpo_ids}
\alias{fix_hpo_ids}
\title{Fix missing HPO IDs}
\usage{
fix_hpo_ids(dat, phenotype_to_genes = load_phenotype_to_genes())
}
\arguments{
\item{dat}{\link[data.table]{data.table}
containing a column named "hpo_name".}
}
\value{
data.table with the column "hpo_id"
}
\description{
Fix missing HPO IDs by inferring their identity from the \code{hpo_name}
column using several methods.
}
\keyword{internal}
