% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/get_hpo.R, R/get_hpo_id_direct.R,
%   R/get_mappings.R
\name{get_}
\alias{get_}
\alias{get_hpo}
\alias{get_hpo_id_direct}
\alias{get_mappings}
\title{Get functions}
\usage{
get_hpo(
  lvl = 2,
  force_new = FALSE,
  terms = NULL,
  tag = "v2024-02-08",
  method = "github",
  save_dir = KGExplorer::cache_dir(),
  ...
)

get_hpo_id_direct(term, hpo = get_hpo())

get_mappings(
  terms = NULL,
  to = c("UMLS", "SNOMED", "ICD9", "ICD10"),
  max_dist = 1,
  force_new = FALSE,
  save_dir = KGExplorer::cache_dir(package = "HPOExplorer"),
  ...
)
}
\arguments{
\item{lvl}{How many levels deep into the ontology to get ancestors from.
For example:
\describe{
\item{1}{"All"}
\item{2}{"Phenotypic abnormality"}
\item{3}{"Abnormality of the nervous system"}
\item{4}{"Abnormality of nervous system physiology"}
\item{5}{"Neurodevelopmental abnormality" or "Behavioral abnormality"}
}}

\item{force_new}{Force a new download or creation of a data resource.}

\item{terms}{A subset of HPO IDs to assign Tiers to.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{method}{Method to construct plot with.}

\item{save_dir}{Directory to save a file to.}

\item{...}{
  Arguments passed on to \code{\link[KGExplorer:get_]{KGExplorer::get_ontology}}, \code{\link[=get_data]{get_data}}
  \describe{
    \item{\code{name}}{\describe{
\item{<...>}{Any ontology name from \link[KGExplorer]{get_ols_options}}
\item{"mondo"}{
Import the \href{https://mondo.monarchinitiative.org/}{Mondo} ontology.
\href{https://github.com/monarch-initiative/mondo/release}{
GitHub.}}.
\item{"hpo"}{
Import the \href{https://hpo.jax.org/app/}{Human Phenotype Ontology}.
\href{https://github.com/obophenotype/human-phenotype-ontology/release}{
GitHub.}}
}}
    \item{\code{filetype}}{File type to search for.}
    \item{\code{import_func}}{Function to import the ontology with. 
If \code{NULL}, automatically tries to choose the correct function.}
    \item{\code{add_metadata}}{Add metadata to the resulting ontology object.}
    \item{\code{add_n_edges}}{Add the number of edges (connections) for each term.}
    \item{\code{add_ontology_levels}}{Add the ontology level for each term.}
    \item{\code{add_version}}{Add the release version
to the returned object's \link[base]{attributes}}
    \item{\code{file}}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}
    \item{\code{repo}}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}
    \item{\code{overwrite}}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
  }}

\item{term}{One or more ontology IDs.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{to}{Medical ontologies to provide mappings to.}

\item{max_dist}{Maximum cross-ontology distance to search for mappings.
A distance of 1 means only direct mappings are returned.
Greater distances mean that mappings are found through intermediate terms
and are therefore less reliable.}
}
\value{
Data.

\link[simona]{ontology_DAG} object.

The HPO ID(s) of phenotype(s)

A named list of \link[data.table]{data.table}
objects containing mappings.
}
\description{
Functions to get data objects and extract elements.
}
\section{Functions}{
\itemize{
\item \code{get_hpo()}: get_
Get Human Phenotype Ontology (HPO)

Downloads and imports the "v2024-02-08" release of the
Human Phenotype Ontology (HPO) for the purposes of version control and
consistency of results. To get the latest version of the HPO,
use \code{get_hpo(tag="release", force_new=T)}
The HPO objects are created from the OBO files distributed via
the official
\href{https://github.com/obophenotype/human-phenotype-ontology}{
HPO GitHub repository}.

By comparison, the \code{hpo} data from \pkg{ontologyIndex} is from 2016.
Note that the maximum ontology level depth in the 2016 version was 14,
whereas in the 2023 version the maximum ontology level depth is 16
(due to an expansion of the HPO).

\item \code{get_hpo_id_direct()}: get_
Get HPO term ID direct

Directly retrieves it from the HPO
ontology object.

\item \code{get_mappings()}: get_
Get Human Phenotype Ontology (HPO) mappings

Get mappings between HPO terms and their corresponding IDs in
other medical ontologies (e.g. SNOMED CT, UMLS, ICD-9/10).

}}
\examples{
hpo <- get_hpo()
term = "Phenotypic abnormality"
pheno_abnormality_id <- get_hpo_id_direct(term = term)
mappings <- get_mappings()
}
\concept{get_}
