% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_description_dataframe.R
\name{get_disease_description_dataframe}
\alias{get_disease_description_dataframe}
\title{Generate HPO disease description dataframe}
\usage{
get_disease_description_dataframe(HPO_ids, verbose = TRUE)
}
\arguments{
\item{HPO_ids}{A character vector of HPO Ids}

\item{verbose}{Print messages.}
}
\value{
A dataframe of HPO phenotype descriptions
}
\description{
This function will interact with the HPO API to create a dataframe that
contains a column of HPO IDs and a column of short descriptions describing
the phenotype.
}
\details{
This can take a while to run on a large number of IDs so it is recommended
that you do it once and save the dataframe.
}
\examples{
genedata <- load_phenotype_to_genes()
HPO_ids <- unique(genedata$ID)[seq_len(5)]
description_df <- get_disease_description_dataframe(HPO_ids)
}
