% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gencc.R
\name{get_gencc}
\alias{get_gencc}
\title{Get GenCC}
\usage{
get_gencc(
  agg_by = c("disease_id", "gene_symbol"),
  dict = c(Definitive = 6, Strong = 5, Moderate = 4, Supportive = 3, Limited = 2,
    `Disputed Evidence` = 1, `Refuted Evidence` = 0, `No Known Disease Relationship` = 0),
  save_dir = tools::R_user_dir("HPOExplorer", which = "cache"),
  verbose = TRUE
)
}
\arguments{
\item{agg_by}{Column to aggregate metadata by.}

\item{dict}{A named vector of evidence score mappings.
See \href{https://thegencc.org/faq.html#validity-termsdelphi-survey}{here}
for more information.}

\item{save_dir}{Directory to save data to.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} with columns:
\itemize{
\item{"disease_id": }{Disease ID.}
\item{"gene_symbol": }{Gene symbol.}
\item{"evidence_score": }{Evidence score.}
}
}
\description{
Get phenotype-gene evidence score from the
\href{https://thegencc.org/}{Gene Curation Coalition}.
Data downloaded from \href{https://search.thegencc.org/download}{here}.\cr
\emph{NOTE:} Due to licensing restrictions, a GenCC download does not
include OMIM data. OMIM data can be accessed and downloaded through
\href{https://omim.org/downloads/}{OMIM}.\cr
\emph{NOTE:} GenCC does not currently have any systematic versioning.
There for the \code{attr(obj,"version")} attribute is set to the date it was
downloaded and cached by \link[HPOExplorer]{get_gencc}.
}
\examples{
d <- get_gencc()
}
