% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_list.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Get gene list}
\usage{
get_gene_list(phenotype, phenotype_to_genes)
}
\arguments{
\item{phenotype}{A phenotype (name) from the HPO
(e.g. "Phenotypic abnormality") <string>}

\item{phenotype_to_genes}{The phenotype_to_genes.txt file from HPO that
contains gene annotations for HPO phenotypes.}
}
\value{
a character vector of genes associated with the selected phenotype.
}
\description{
This function returns a vector of genes associated with a particular phenotype.
It uses the \code{phenotype_to_genes.txt} file from the hpo.
}
