% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_lists.R
\name{get_gene_lists}
\alias{get_gene_lists}
\title{Get gene list}
\usage{
get_gene_lists(
  phenotypes,
  phenotype_to_genes = load_phenotype_to_genes(),
  hpo = get_hpo(),
  as_list = FALSE,
  hpo_id_labels = TRUE
)
}
\arguments{
\item{phenotypes}{One or more phenotype name from the HPO
(e.g. "Phenotypic abnormality") \<string\>.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{as_list}{Return as a named list instead of a data.table.}

\item{hpo_id_labels}{Label the gene lists with HPO IDs
(instead of phenotype names).}
}
\value{
a character vector of genes associated with the selected phenotype.
}
\description{
This function returns a vector of genes associated with a
particular phenotype.
It uses the \code{phenotype_to_genes.txt} file from the HPO.
}
\examples{
phenotypes <- c("Focal motor seizure","HP:0000002","HP:0000003")
gene_lists <- get_gene_lists(phenotypes = phenotypes)
}
