% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hierarchy.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Get component relative ontology level of all terms within a subset of HPO}
\usage{
get_hierarchy(adjacency, hpo, reverse = TRUE)
}
\arguments{
\item{adjacency}{A adjacency matrix of phenotypes where 1 represents i is parent of j
and 0 represents that i is not a parent of j. It is a subset of the main phenotype
adjacency matrix.}

\item{hpo}{The HPO ontology data object}

\item{reverse}{A boolean, if TRUE it will reverse the ontology level numbers so that
the parent terms are larger than the child terms.}
}
\value{
A named vector of relative ontology level, where names are HPO Ids and
value is relative ontology level.
}
\description{
This calls the \code{find_parent} function on all phenotypes in the subset of
the HPO to be plotted. The subest chosen when creating the adjacency from the main
adjacency matrix of all phenotypes. So, the phenotypes to be plotted can be
found in the row and column names of adjacency.
}
\keyword{internal}
