% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpo_id.R
\name{get_hpo_id}
\alias{get_hpo_id}
\title{Get HPO ID from phenotype name}
\usage{
get_hpo_id(phenotype, phenotype_to_genes)
}
\arguments{
\item{phenotype}{Phenotype name from the HPO \<string\>}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}
}
\value{
The HPO Id \<string\>
}
\description{
I have done this more efficiently elsewhere using the HPO data object.
May be worth replacing, or just add the HPO Id to all data points
in the results permanently.
}
\examples{
genedata <- load_phenotype_to_genes()
phenotype <- "Phenotypic abnormality"
pheno_abnorm_id <- get_hpo_id(phenotype, genedata)
}
