% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpo_id_direct.R
\name{get_hpo_id_direct}
\alias{get_hpo_id_direct}
\title{Get HPO term ID direct}
\usage{
get_hpo_id_direct(phenotype = "Phenotypic abnormality", hpo = get_hpo())
}
\arguments{
\item{phenotype}{The phenotype of interest \<string\>}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}
}
\value{
The HPO ID of phenotype \<string\>
}
\description{
This is similar to the \code{get_hpo_germID} function except it does not take
the ID from the phenotype_to_genes dataframe.
It directly retrieves it from the HPO
ontology object (found in the ontologyIndex package \code{data(hpo)}). This
allows you to get the ID of terms that do not have gene annotations.
}
\details{
Why not always use this? I have found that occasionally it doesn't work.
Also, it requires you to load the full HPO object to get an ID,
which is wasteful if you are only using the gene annotations.
}
\examples{
pheno_abnormality_id <- get_hpo_id_direct("Phenotypic abnormality")
}
