% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kg_monarch.R
\name{get_kg_monarch}
\alias{get_kg_monarch}
\title{Get knowledge graph: Monarch}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/37707514/}{BioThings Explorer}
}
\usage{
get_kg_monarch()
}
\description{
Option 1: Use the \href{https://api.monarchinitiative.org/api/}{biolink API}
to efficiently extract specific subset of data
from the Monarch server.
Option 2: Import the entire knowledge graph from the
\href{https://data.monarchinitiative.org/monarch-kg/latest/}{Monarch server}.
}
