% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mondo.R
\name{get_mondo}
\alias{get_mondo}
\title{Get Mondo Disease Ontology}
\source{
\code{
mondo <- HPOExplorer:::make_ontology(file="mondo-base.obo",
                       repo="monarch-initiative/mondo",
                       upload_tag="latest")
}
}
\usage{
get_mondo(
  save_dir = tools::R_user_dir("HPOExplorer", which = "cache"),
  tag = "latest",
  overwrite = TRUE
)
}
\arguments{
\item{save_dir}{Directory to save data to.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
}
\value{
\link[ontologyIndex]{ontology_index}
}
\description{
Import \href{https://mondo.monarchinitiative.org/}{Mondo} as an R object.
}
\examples{
mondo <- get_mondo()
}
