% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ont_lvls.R
\name{get_ont_lvls}
\alias{get_ont_lvls}
\title{Get ontology level for HPO terms}
\usage{
get_ont_lvls(
  terms,
  hpo = get_hpo(),
  adjacency = NULL,
  absolute = TRUE,
  exclude_top_lvl = TRUE,
  reverse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{terms}{Character vector of HPO IDs.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{An adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{absolute}{Make the levels absolute in the sense that they consider
the entire HPO ontology (\code{TRUE}).
Otherwise, levels will be relative to only the HPO terms that are in
the provided subset of \code{terms} AND are directly adjacent (connected)
to a given cluster of terms (\code{FALSE}).}

\item{exclude_top_lvl}{Exclude the top level term of the HPO
(i.e. "All" (HP:0000001)) when computing \emph{absolute} levels.
This argument is not used when computing \emph{relative} levels.}

\item{reverse}{If \code{TRUE}, ontology
level numbers with be revered such that the level of the parent terms
are larger than the child terms.}

\item{verbose}{Print messages.}
}
\value{
A named vector of relative ontology level,
where names are HPO Ids and
value is relative ontology level.
}
\description{
For a given set of HPO terms, get their level
within the hierarchically organised HPO ontology.
Ontology level can be computed either absolute mode (\code{absolute=TRUE})
where the entire ontology is considered when assigning levels, or
relative mode (\code{absolute=FALSE}) where only a subset of the ontology
that is connected to a given term is considered when assigning levels.
Relative mode can be helpful when trying to make plot where nodes are
scaled to the ontology level.
}
\examples{
terms <- ontologyIndex::get_descendants(ontology = get_hpo(),
                                        roots = "HP:0000002")
lvls <- get_ont_lvls(terms = terms)
lvls_abs <- get_ont_lvls(terms = terms, absolute=TRUE)
}
