% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ont_lvls2.R
\name{get_ont_lvls2}
\alias{get_ont_lvls2}
\title{Get relative ontology level for multiple HPO terms}
\usage{
get_ont_lvls2(
  terms,
  hpo = get_hpo(),
  adjacency = adjacency_matrix(terms = terms, hpo = hpo),
  absolute = FALSE,
  reverse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{terms}{Character vector of HPO IDs.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{AN adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{absolute}{Make the levels absolute in the sense that they consider
the entire HPO ontology (\code{TRUE}).
Otherwise, levels will be relative to only the subset \code{terms} provided
(\code{FALSE}).}

\item{reverse}{A boolean, if TRUE it will reverse the ontology
level numbers so that
the parent terms are larger than the child terms.}

\item{verbose}{Print messages.}
}
\value{
A named vector of relative ontology level,
where names are HPO Ids and
value is relative ontology level.
}
\description{
This calls the \code{get_ont_lvl} function on all phenotypes in a
subset of the HPO. The subset chosen when creating the adjacency from the main
adjacency matrix of all phenotypes. So, the phenotypes can be found in the
row and column names of adjacency.
}
\examples{
terms <- ontologyIndex::get_descendants(ontology = get_hpo(),
                                        roots = "HP:0000002")
lvls <- get_ont_lvls2(terms = terms)
}
