% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relative_ont_level_multiple.R
\name{get_relative_ont_level_multiple}
\alias{get_relative_ont_level_multiple}
\title{Get relative ontology level for multiple HPO terms}
\usage{
get_relative_ont_level_multiple(phenoAdj, hpo, reverse = TRUE)
}
\arguments{
\item{phenoAdj}{A adjacency matrix of phenotypes where 1 represents i is parent of j
and 0 represents that i is not a parent of j. It is a subset of the main phenotype adjacency matrix}

\item{hpo}{The HPO ontology data object}

\item{reverse}{A boolean, if TRUE it will reverse the ontology level numbers so that
the parent terms are larger than the child terms.}
}
\value{
A named vector of relative ontology level, where names are HPO Ids and
value is relative ontology level.
}
\description{
This calls the \code{get_relative_ont_level} function on all phenotypes in a
subset of the HPO. The subest chosen when creating the phenoAdj from the main
adjacency matrix of all phenotypes. So, the phenotypes can be found in the
row and column names of phenoAdj.
}
