% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_term_definition_api.R
\name{get_term_definition_api}
\alias{get_term_definition_api}
\title{Get term definition: API}
\usage{
get_term_definition_api(term, line_length = FALSE)
}
\arguments{
\item{term}{One or more HPO IDs.}
}
\value{
Character vector of definitions.
}
\description{
This function accesses the HPO API to get a description/definition of an
HPO term. If a \code{line_length} \> 0 is passed to the function, it will add
newlines every nth word. This can be useful when displaying the description
in plots with limited space.
}
\keyword{internal}
