% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upheno.R
\name{get_upheno}
\alias{get_upheno}
\title{Get UPHENO}
\usage{
get_upheno(file = c("ontology", "bestmatches", "upheno_mapping"))
}
\arguments{
\item{file}{Can be one of the following:
\itemize{
\item{"ontology"}{Creates an \link[ontologyIndex]{ontologyIndex} R object by
importing the OBO file directly from the official
\href{https://github.com/obophenotype/upheno}{UPHENO GitHub repository}.}
\item{"bestmatches"}{Returns a merged table with the best matches between
human and non-human homologous phenotypes (from multiple species).
Distributed by the official
\href{https://github.com/obophenotype/upheno/tree/master/mappings}{
UPHENO GitHub repository}.}
\item{"upheno_mapping"}{Return a merged table with matches bteween human
and non-human homologous phenotypes (from multiple species).
Distributed by the
\href{https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html}{
Monarch Initiative server}.}
}}
}
\value{
\link[ontologyIndex]{ontologyIndex} or
\link[data.table]{data.table}.
}
\description{
Get data from the \href{https://github.com/obophenotype/upheno}{
Unified Phenotype Ontology (UPHENO)}.
}
\examples{
upheno <- get_upheno()
}
