% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{Get version}
\usage{
get_version(obj, return_version = FALSE, verbose = TRUE)
}
\arguments{
\item{obj}{An object from
\link[HPOExplorer]{get_hpo} or
\link[HPOExplorer]{load_phenotype_to_genes}.}

\item{return_version}{Return the version as a character string.}

\item{verbose}{Print messages.}
}
\value{
HPO Release version a character string.
}
\description{
Extract the precise version of the HPO release that the data object
was built from. All HPO Releases can be found at the official
\href{https://github.com/obophenotype/human-phenotype-ontology/releases}{
HPO GitHub Releases page}.
}
\examples{
obj <- get_hpo()
get_version(obj=obj)
}
