% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_check.R
\name{gpt_annot_check}
\alias{gpt_annot_check}
\title{Check annotations from GPT}
\usage{
gpt_annot_check(
  annot = gpt_annot_read(),
  query_hits = search_hpo(),
  verbose = TRUE
)
}
\arguments{
\item{annot}{GPT-generated phenotype annotations.}

\item{query_hits}{A named list of HPO IDs for each query,
generated by \link[HPOExplorer]{search_hpo}.}

\item{verbose}{Print messages.}
}
\value{
Named list
}
\description{
Check GPT phenotype annotations using a several metrics.
}
\examples{
checks <- gpt_annot_check()
}
