% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_class.R
\name{gpt_annot_class}
\alias{gpt_annot_class}
\title{GPT annotations: add severity class}
\usage{
gpt_annot_class(
  res_coded = gpt_annot_codify(),
  inclusion_values = c(2, 3),
  tiers_dict = list(death = 1, intellectual_disability = 1, impaired_mobility = 2,
    physical_malformations = 2, blindness = 3, sensory_impairments = 3, immunodeficiency
    = 3, cancer = 3, reduced_fertility = 4),
  add_severity_score_gpt = TRUE
)
}
\arguments{
\item{res_coded}{The output of list \link{gpt_annot_codify}.}

\item{inclusion_values}{Numerically encoded annotation values
to count as hits. See the \code{gpt_annot_codify(code_dict=)}
argument defaults for the mappings.
Defaults to 2 ("often") and 3 ("always").}

\item{tiers_dict}{A named list of severity tiers.}

\item{add_severity_score_gpt}{Whether to add a continuous severity score
as well.}
}
\value{
A data.table with severity classes, as well as severity
scores (0-100) which can be used to rank severity within each class.
}
\description{
Convert severity annotations ot severity classes
(profound > severe > moderate > mild) to approximate those introduced in
\href{https://doi.org/10.1371/journal.pone.0114391}{Lazarin et al. 2014}.
}
\examples{
res_coded <- gpt_annot_codify()
res_class <- gpt_annot_class(res_coded)
}
