% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_plot.R
\name{gpt_annot_plot}
\alias{gpt_annot_plot}
\title{Plot annotations from GPT}
\usage{
gpt_annot_plot(
  annot = gpt_annot_read(),
  keep_ont_levels = seq(0, 5),
  remove_descendants = c("Clinical course", "Sporadic", "Multifactorial inheritance",
    "Inheritance modifier", "Phenotypic variability"),
  top_n = 50,
  verbose = TRUE
)
}
\arguments{
\item{annot}{GPT-generated phenotype annotations.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link{add_ont_lvl} for details.}

\item{remove_descendants}{Remove HPO terms that are descendants of a given
ancestral HPO term. Ancestral terms be provided as a character vector of
phenotype names (e.g. \code{c("Clinical course")}),
HPO IDs (e.g. \code{"HP:0031797" }) or a mixture of the two.
See \link{add_ancestor} for details.}

\item{top_n}{Top number of most severe phenotypes to plot in heatmap.}

\item{verbose}{Print messages.}
}
\value{
Named list of plots.
}
\description{
Plot annotations from GPT.
}
\examples{
plots <- gpt_annot_plot()
}
