% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_plot_branches.R
\name{gpt_annot_plot_branches}
\alias{gpt_annot_plot_branches}
\title{Plot annotations from GPT: by branch}
\usage{
gpt_annot_plot_branches(
  hpo = get_hpo(),
  gpt_annot = gpt_annot_read(hpo = hpo),
  keep_descendants = NULL,
  keep_ont_levels = NULL,
  metric = "congenital_onset",
  fill_lab = gsub("_", " ", metric),
  show_plot = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{gpt_annot}{Output from \link{gpt_annot_read}.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link{add_ont_lvl} for details.}

\item{metric}{Annotation metric to plot.}

\item{fill_lab}{Fill label in legend.}

\item{show_plot}{Show the plot.}
}
\value{
Named list of plot and data.
}
\description{
Plot annotations from GPT by ancestral HPO branch.
}
\examples{
out <- gpt_annot_plot_branches()
}
