% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_phenotypes.R
\name{harmonise_phenotypes}
\alias{harmonise_phenotypes}
\title{Harmonise phenotypes}
\usage{
harmonise_phenotypes(
  phenotypes,
  hpo = get_hpo(),
  as_hpo_ids = FALSE,
  ignore_case = TRUE
)
}
\arguments{
\item{phenotypes}{A character vector of phenotype names and/or HPO IDs.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{as_hpo_ids}{Return all \code{phenotypes} as HPO IDs (\code{TRUE}),
or phenotype names (\code{FALSE}).}

\item{ignore_case}{Ignore casing when searching for matches.}
}
\value{
Character vector
}
\description{
Harmonise a mixed vector of phenotype names (e.g. "Focal motor seizures")
and HPO IDs (e.g. c("HP:0000002","HP:0000003")).
}
\examples{
phenotypes <- c("Focal motor seizures","HP:0000002","HP:0000003")
#### As phenotype names ####
pheno_names <- harmonise_phenotypes(phenotypes=phenotypes)
#### As HPO IDs ####
pheno_ids <- harmonise_phenotypes(phenotypes=phenotypes, as_hpo_ids=TRUE)
}
