% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo}
\alias{hpo}
\title{Human Phenotype Ontology (HPO)}
\format{
ontology_index
}
\source{
\href{https://bioportal.bioontology.org/ontologies/HP}{BioPortal}

\code{
URL <- paste0(
  "https://data.bioontology.org/ontologies/HP/submissions/599/download",
  "?apikey=8b5b7825-538d-40e0-9e9e-5ab9274a9aeb"
)
tmp <- tempfile(fileext = "hpo.obo")
utils::download.file(URL, tmp)
hpo <- ontologyIndex::get_OBO(tmp)
#' ### Fix non-ASCII characters in metadata ####
func <- function(v){
  Encoding(v) <- "latin1"
  iconv(v, "latin1", "UTF-8")
}
attributes(hpo)$version <- func(attributes(hpo)$version)
usethis::use_data(hpo, overwrite = TRUE)
}
}
\usage{
data("hpo")
}
\description{
Updated version of Human Phenotype Ontology (HPO) from 2023-01-27.
Created from the OBO files distributed by
\href{https://bioportal.bioontology.org/ontologies/HP}{BioPortal}.
By comparison, the \code{hpo} data from \pkg{ontologyIndex} is from 2016.
Note that the maximum ontology level depth in the 2016 version was 14,
whereas in the 2023 version the maximum ontology level depth is 16
(due to an expansion of the HPO).
}
\keyword{datasets}
