% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_api.R
\name{hpo_api}
\alias{hpo_api}
\title{HPO API}
\usage{
hpo_api(
  hpo_id,
  type = list(NULL, "genes", "diseases")[[1]],
  url = paste("hpo.jax.org/api/hpo/term", hpo_id, type, sep = "/")
)
}
\arguments{
\item{hpo_id}{HPO ID for a given term.}

\item{type}{Type of data to extract. Set \code{NULL} to return all data.}

\item{url}{API entry point.
See \href{https://hpo.jax.org/api/hpo/docs}{HPO API documentation}
for details.}
}
\value{
A named list or data.frame of metadata for a given HPO ID.
}
\description{
Extract data from the Human Phenotype Ontology (HPO) using the
Application Programming Interface (API).
}
\examples{
dat <- hpo_api(hpo_id="HP:0011420", type="diseases")
}
