% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_death}
\alias{hpo_death}
\title{Human Phenotype Ontology: Age of Death}
\format{
data.table
}
\source{
\code{
terms <- ontologyIndex::get_descendants(ontology = hpo,
                                        roots = "HP:0011420",
                                        exclude_roots = TRUE)
aod <- lapply(stats::setNames(terms, terms),
              function(hpo_id){
                messager("Extracting API data for",hpo_id,v=verbose)
                d <- hpo_api(hpo_id = hpo_id, type = "diseases")$diseases
              }) |> data.table::rbindlist(fill = TRUE,
                                          use.names = TRUE,
                                          idcol = "AgeOfDeath")
aod$AgeOfDeath_name <- harmonise_phenotypes(phenotypes = aod$AgeOfDeath,
                                            as_hpo_ids = FALSE)
#### Convert AoD to numeric scores ####
dict <- hpo_dict(type="AgeOfDeath")
aod$AgeOfDeath_score <- dict[aod$AgeOfDeath_name]
#### Link diseases to HPO phenotypes ####
annot <- load_phenotype_to_genes(filename = "phenotype.hpoa")
hpo_death <- data.table::merge.data.table(aod,
                                          annot[,c("HPO_ID","DiseaseName")],
                                          by.x="diseaseName",
                                          by.y = "DiseaseName")
usethis::use_data(hpo_death, overwrite = TRUE)
}
}
\usage{
data("hpo_death")
}
\description{
Age of Death associated with each disease, and by extension, each phenotype.
}
\keyword{datasets}
