% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_meta}
\alias{hpo_meta}
\title{Human Phenotype Ontology: metadata}
\format{
data.table
}
\source{
\href{https://bioportal.bioontology.org/ontologies/HP}{BioPortal}

\code{
URL <- "https://data.bioontology.org/ontologies/HP/download?apikey=8b5b7825-538d-40e0-9e9e-5ab9274a9aeb&download_format=csv"
tmp <- tempfile(fileext = "hpo.csv.gz")
utils::download.file(URL, tmp)
hp <- data.table::fread(tmp)
hp <- hp[,hpo_id:=basename(hp$`Class ID`)][,link:=hp$`Class ID`][Obsolete==FALSE]
cols <- c("hpo_id","link","Preferred Label","Synonyms",#"Definitions",
          "CUI","database_cross_reference","definition")
hpo_meta <- hp[,cols, with=FALSE]
names(hpo_meta) <- gsub(" +","",tolower(names(hpo_meta)))
### Make ASCI-friendly
hpo_meta <- HPOExplorer:::as_ascii(dt=hpo_meta)
hpo_meta$hpo_id=gsub("_",":",hpo_meta$hpo_id)
data.table::setnames(hpo_meta,"hpo_id","HPO_ID")
for(nm in c("UMLS","SNOMEDCT_US","MSH")){
  hpo_meta[[paste0(nm,"_ID")]] <- lapply(
    strsplit(hpo_meta$database_cross_reference,"\\|"),
                                     function(x){
    s <- grep(paste0(nm,":"),x,value = TRUE)
    if(length(s)==0)NA else s
                                     })
}
usethis::use_data(hpo_meta, overwrite = TRUE)
}
}
\usage{
data("hpo_meta")
}
\description{
Additional metadata from the \href{https://hpo.jax.org/}{HPO}.
}
\keyword{datasets}
