% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_onsets}
\alias{hpo_onsets}
\title{Human Phenotype Ontology: disease onset}
\format{
data.table
}
\source{
\code{
annot <- load_phenotype_to_genes(filename = "phenotype.hpoa")
annot <- annot[onset!="",]
annot$onset_name <- harmonise_phenotypes(phenotypes = annot$onset,
                                         as_hpo_ids = FALSE)
counts <- dplyr::group_by(annot, disease_id) |>
  dplyr::summarise(hpo_ids=length(unique(hpo_id)),
                   onsets=length(unique(onset)))
## The number of onsets partially depends on the number of hpo_ids
## so it's necessary to keep hpo_id too.
cor(counts$hpo_ids, counts$onsets)
hpo_onsets <- annot[,c("disease_id","hpo_id","onset","onset_name")]
usethis::use_data(hpo_onsets, overwrite = TRUE)
}
}
\usage{
data("hpo_onsets")
}
\description{
HPO disease onsets.
}
\keyword{datasets}
