% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_term_definition_list.R
\name{hpo_term_definition_list}
\alias{hpo_term_definition_list}
\title{Get HPO term description for a list of terms}
\usage{
hpo_term_definition_list(ontologyId_list, disease_descriptions)
}
\arguments{
\item{ontologyId_list}{A character vector of HPO Ids.}

\item{disease_descriptions}{A data frame of disease descriptions
for all HPO Id.}
}
\value{
A named vector of disease descriptions,
with HPO Id as names and descriptions as values.
}
\description{
Need to redo this without a for loop (use \code{lapply} or something).
This just applies the \link[HPOExplorer]{hpo_get_term_definition} function
to a character vector of terms.
}
\examples{
genedata <- load_phenotype_to_genes()
HPO_ids <- unique(genedata$ID[1:10])

disease_descriptions <- get_disease_description_dataframe(HPO_ids)
hpo_term_definition_list(HPO_ids[1:3], disease_descriptions)
}
