% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_tiers}
\alias{hpo_tiers}
\title{Human Phenotype Ontology: disease severity tiers}
\format{
data.table
}
\source{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4262393/}{
Lazarin et al (2014)}

\code{
file <- system.file("extdata","hpo_tiers.csv.gz",package = "HPOExplorer")
hpo_tiers <- data.table::fread(file)
usethis::use_data(hpo_tiers, overwrite = TRUE)
}
}
\usage{
data("hpo_tiers")
}
\description{
A list of the 49 HPO disease phenotypes in 4 different tiers
from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4262393/}{
Lazarin et al (2014)}, specifically Table 1.
HPO IDs of all the terms are displayed in the right column.
}
\keyword{datasets}
