% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_to_matrix.R
\name{hpo_to_matrix}
\alias{hpo_to_matrix}
\title{HPO to matrix}
\usage{
hpo_to_matrix(
  terms = NULL,
  phenotype_to_genes = load_phenotype_to_genes(),
  formula = "gene_symbol ~ hpo_name",
  fun.aggregate = mean,
  value.var = "evidence_score_mean",
  fill = 0,
  run_cor = FALSE,
  as_matrix = TRUE,
  as_sparse = TRUE,
  method = "pearson",
  verbose = TRUE
)
}
\arguments{
\item{terms}{A subset of HPO IDs to include.
Set to \code{NULL} (default) to include all terms.}

\item{phenotype_to_genes}{Output of
\link{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{formula}{A formula of the form LHS ~ RHS to cast, see Details.}

\item{fun.aggregate}{Should the data be aggregated before casting? If the formula doesn't identify a single observation for each cell, then aggregation defaults to \code{length} with a message.

  To use multiple aggregation functions, pass a \code{list}; see Examples. }

\item{value.var}{Name of the column whose values will be filled to cast. Function \code{guess()} tries to, well, guess this column automatically, if none is provided.

  Cast multiple \code{value.var} columns simultaneously by passing their names as a \code{character} vector. See Examples. }

\item{fill}{Value with which to fill missing cells. If \code{fun.aggregate} is present, takes the value by applying the function on a 0-length vector.}

\item{run_cor}{Return a matrix of pairwise correlations.}

\item{as_matrix}{Return the results as a matrix (\code{TRUE}).
Otherwise, will return the results as a \link[data.table]{data.table}
with an extra column "gene_symbol".}

\item{as_sparse}{Convert the data to a sparse matrix.
Only used when \code{as_matrix=TRUE}.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{verbose}{Print messages.}
}
\value{
A gene x phenotype matrix,
or a phenotype x phenotype matrix if \code{run_cor=TRUE}.
}
\description{
Convert gene-phenotype associations from the Human Phenotype Ontology (HPO)
into a gene x phenotype matrix. The returned matrix is sparse and binary,
such that 1 indicates a the gene is associated with a given phenotype
according to the HPO annotation, and 0 indicates it is not.
By default, full phenotype names are used as the column names
(e.g. "Abnormality of body height"),
however you can instead set them to the HPO IDs
by changing the \code{formula} argument to:
\code{formula = "gene_symbol ~ hpo_id"}.
Phenotypes that are not present in the \code{phenotype_to_genes} annotations
are omitted from the final matrix.
}
\examples{
phenos <- example_phenos()
X <- hpo_to_matrix(terms = phenos$hpo_id)
}
