% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_to_plotly.R
\name{igraph_to_plotly}
\alias{igraph_to_plotly}
\title{igraph to plotly data}
\usage{
igraph_to_plotly(
  g,
  layout_func = igraph::layout.fruchterman.reingold,
  dim = 3,
  seed = 2023,
  verbose = TRUE
)
}
\arguments{
\item{g}{\link[igraph]{igraph} object
generated by \link[HPOExplorer]{make_igraph_object}.}

\item{layout_func}{Layout function for the graph.}

\item{dim}{Number of dimensions to create layout in.}

\item{seed}{Random seed to enable reproducibility.}

\item{verbose}{Print messages.}
}
\value{
Named list of data.frames.
}
\description{
Convert an igraph to data for a 3D plotly plot.
}
\keyword{internal}
