% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_surface.R
\name{kde_surface}
\alias{kde_surface}
\title{Kernel Density Estimation (KDE) surface}
\usage{
kde_surface(xyz, n = 50, extend_kde = 1, rescale_z = TRUE)
}
\arguments{
\item{xyz}{A data.frame containing the columns "x" and "y" (required).
Can also include a third column "z" (optional).}

\item{n}{
Number of grid points in each direction.  Can be scalar or a length-2
integer vector.
}

\item{extend_kde}{Extend the limits of the KDE area by setting
\code{extend_kde>1}.
This makes the "mountains' of the KDE end more smoothly along the edges of
the x/y coordinates.}

\item{rescale_z}{Rescale the z-axis so that the KDE can be positioned
below the original data points (when plotting them together).}
}
\value{
\itemize{
\item{\code{x, y}: }{
The x and y coordinates of the grid points, vectors of length n.
}
\item{\code{z}: }{
An \code{n[1]} by \code{n[2]} matrix of the estimated density:
rows correspond to the value of \code{x}, columns to the value of \code{y}.
}
}
}
\description{
Compute a KDE surface based on the density of points along
an x- and y-axis.
}
\examples{
xyz <- data.frame(x=stats::rnorm(50),
                  y=stats::rnorm(50),
                  z=stats::rnorm(50))
kd <- kde_surface(xyz = xyz)
}
