% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_columns.R
\name{list_columns}
\alias{list_columns}
\title{List columns}
\usage{
list_columns(columns = NULL, add_gene_cols = FALSE, extra_cols = NULL)
}
\arguments{
\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[KGExplorer]{add_hoverboxes}.}

\item{add_gene_cols}{Add columns related to genes.}

\item{extra_cols}{Additional user-specified column names.}
}
\value{
Named list of columns.
}
\description{
Provide a list of columns to be used as hover metadata in an interactive
network plot generated by \link[HPOExplorer]{make_network_plot}.
}
\examples{
columns <- list_columns()
}
