% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_onsets.R
\name{list_onsets}
\alias{list_onsets}
\title{List onsets}
\usage{
list_onsets(
  hpo = get_hpo(),
  postnatal_only = FALSE,
  as_hpo_ids = TRUE,
  include_na = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{postnatal_only}{Only include postnatal onsets.}

\item{as_hpo_ids}{Return as a character vector vector HPO IDs only.}

\item{include_na}{Include NA values for Onset.}

\item{verbose}{Print messages.}
}
\value{
Named list of HPO IDs.
}
\description{
List Onset phenotypes in the HPO.
}
\examples{
onsets <- list_onsets()
}
