% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_disease_genes.R
\name{load_disease_genes}
\alias{load_disease_genes}
\title{Load disease genes}
\usage{
load_disease_genes(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\value{
data.table
}
\description{
Load gene lists associated with each disease phenotype from:
\itemize{
\item{DECIPHER}
\item{ORPHANET}
\item{OMIM}
}
}
\examples{
dat <- load_disease_genes()
}
