% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_phenotype_to_genes.R
\name{load_phenotype_to_genes}
\alias{load_phenotype_to_genes}
\title{Load phenotype_to_genes}
\source{
\href{https://hpo.jax.org/app/data/annotations}{HPO source data}
}
\usage{
load_phenotype_to_genes(
  file = c("phenotype_to_genes.txt", "genes_to_phenotype.txt", "phenotype.hpoa"),
  save_dir = file.path(KGExplorer::cache_dir(package = "HPOExplorer"), "data"),
  repo = paste("obophenotype", "human-phenotype-ontology", sep = "/"),
  tag = "latest",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{save_dir}{Folder where the phenotype to genes text
file is/will be stored.}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table} of the HPO annotations.
}
\description{
This is a function for loading the HPO phenotype to genes text file from HPO.
It adds the column names and returns as a dataframe.
It contains phenotypes annotated with associated genes.
}
\examples{
phenotype_to_genes <- load_phenotype_to_genes()
}
