% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/annotate_phenos.R,
%   R/example_phenos.R
\name{main}
\alias{main}
\alias{annotate_phenos}
\alias{example_phenos}
\title{Main functions}
\usage{
annotate_phenos(
  phenos,
  hpo = get_hpo(),
  add_ont_lvl_absolute = TRUE,
  add_ont_lvl_relative = FALSE,
  add_info_content = TRUE,
  add_description = TRUE,
  add_disease_data = FALSE,
  add_ndiseases = add_disease_data,
  add_onsets = add_disease_data,
  add_deaths = add_disease_data,
  add_pheno_frequencies = add_disease_data,
  add_tiers = add_disease_data,
  add_severities = add_disease_data,
  add_disease_definitions = add_disease_data,
  include_mondo = FALSE,
  add_hoverboxes = FALSE,
  columns = list_columns(),
  interactive = TRUE
)

example_phenos(i = seq(10), hpo = get_hpo())
}
\arguments{
\item{phenos}{A data.table containing HPO IDs and other metadata.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{add_ont_lvl_absolute}{Add the absolute ontology level of each HPO term.
See \link[KGExplorer]{get_ontology_levels} for more details.}

\item{add_ont_lvl_relative}{Add the relative ontology level of each HPO term.
See \link[KGExplorer]{get_ontology_levels} for more details.}

\item{add_info_content}{Add information content column for each phenotype.}

\item{add_description}{Whether to get the phenotype descriptions as
well (slower).}

\item{add_disease_data}{Add all disease metadata columns.
This will expand the data using \code{allow.cartesian=TRUE}.}

\item{add_ndiseases}{Add the number of diseases per phenotype.}

\item{add_onsets}{Add age of onset columns using
\link[HPOExplorer]{add_onset}.}

\item{add_deaths}{Add age of death columns using
\link[HPOExplorer]{add_death}.}

\item{add_pheno_frequencies}{Add the frequency of each phenotype in
each disease.}

\item{add_tiers}{Add severity Tiers column using
\link[HPOExplorer]{add_tier}.}

\item{add_severities}{Add severity column using
\link[HPOExplorer]{add_severity}.}

\item{add_disease_definitions}{Add disease definitions.}

\item{include_mondo}{Add \href{https://mondo.monarchinitiative.org/}{MONDO}
IDs, names, and definitions to each disease.}

\item{add_hoverboxes}{Add hoverdata with
\link[KGExplorer]{add_hoverboxes}.}

\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[KGExplorer]{add_hoverboxes}.}

\item{interactive}{Make the plot interactive.}

\item{i}{Indices of HPO IDs to use.}
}
\value{
Data.

phenotype data.table
}
\description{
Main functions.
}
\section{Functions}{
\itemize{
\item \code{annotate_phenos()}: main
Annotate phenotypes

Annotate phenotypes \link[data.table]{data.table} without various types
of metadata.

\item \code{example_phenos()}: main
Example phenotypes dataframe

Create a minimal example of a phenos dataframe.

}}
\examples{
phenos <- example_phenos()
phenos2 <- annotate_phenos(phenos)
phenos <- example_phenos()
}
\concept{main}
