% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/make_igraph_object.R,
%   R/make_network_object.R, R/make_network_plot.R, R/make_phenos_dataframe.R,
%   R/make_tiers.R
\name{make_}
\alias{make_}
\alias{make_igraph_object}
\alias{make_network_object}
\alias{make_network_plot}
\alias{make_phenos_dataframe}
\alias{make_tiers}
\title{Make functions}
\usage{
make_igraph_object(
  phenos,
  hpo = get_hpo(),
  colour_var = "fold_change",
  cols = list_columns(extra_cols = c(colour_var, grep("_count$|_values$", names(phenos),
    value = TRUE))),
  ...
)

make_network_object(
  phenos,
  hpo = get_hpo(),
  colour_var = "fold_change",
  add_ont_lvl_absolute = FALSE,
  cols = list_columns(extra_cols = c(colour_var, grep("_count$|_values$", names(phenos),
    value = TRUE))),
  as = c("ggnetwork", "tbl_graph"),
  ...
)

make_network_plot(
  phenos,
  colour_var = "ontLvl",
  size_var = colour_var,
  label_var = "hpo_name",
  interactive = TRUE,
  show_plot = TRUE,
  hoverbox_column = "hover",
  preferred_palettes = "brewer.reds",
  method = c("ggnetwork", "visnetwork"),
  ...
)

make_phenos_dataframe(
  ancestor = NULL,
  hpo = get_hpo(),
  phenotype_to_genes = load_phenotype_to_genes(),
  add_ont_lvl_absolute = TRUE,
  add_ont_lvl_relative = FALSE,
  add_info_content = FALSE,
  add_description = TRUE,
  add_disease_data = FALSE,
  add_ndiseases = add_disease_data,
  add_onsets = add_disease_data,
  add_deaths = add_disease_data,
  add_pheno_frequencies = add_disease_data,
  add_tiers = add_disease_data,
  add_severities = add_disease_data,
  add_disease_definitions = add_disease_data,
  include_mondo = FALSE,
  add_hoverboxes = FALSE,
  columns = list_columns(),
  interactive = TRUE,
  verbose = TRUE
)

make_tiers(
  hpo = get_hpo(),
  terms = hpo@terms,
  keyword_sets = list(Tier1 = c("intellectual disability", "death"), Tier2 =
    c("impaired mobility", "malform"), Tier3 = c("sight", "vision", "immunodeficien",
    "cancer", "hearing", "touch", "pain ", " pain", "mental illness", "dysmorphic"),
    Tier4 = c("fertility")),
  check_names = TRUE,
  check_definitions = TRUE,
  check_diseases = FALSE,
  search_ancestors = TRUE,
  search_descendants = TRUE,
  as_datatable = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{A data.table containing HPO IDs and other metadata.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{colour_var}{The column from phenos that you wish
to map to node colour.}

\item{cols}{Columns to add to metadata of \link[ggnetwork]{ggnetwork} object.}

\item{...}{
  Arguments passed on to \code{\link[=make_network_object]{make_network_object}}
  \describe{
    \item{\code{}}{}
  }}

\item{add_ont_lvl_absolute}{Add the absolute ontology level of each HPO term.
See \link[KGExplorer]{get_ontology_levels} for more details.}

\item{as}{R object class to return output as.}

\item{size_var}{Column to scale node size by.}

\item{label_var}{Column containing the label for each node in a graph
(e.g. "hpo_name").}

\item{interactive}{Make the plot interactive.}

\item{show_plot}{Show the plot.}

\item{hoverbox_column}{Name of the new hoverbox column to add.}

\item{preferred_palettes}{Preferred palettes to use for each column.}

\item{method}{Method to construct plot with.}

\item{ancestor}{The ancestor to get all descendants of. If \code{NULL},
returns the entirely ontology.}

\item{phenotype_to_genes}{Output of
\link{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{add_ont_lvl_relative}{Add the relative ontology level of each HPO term.
See \link[KGExplorer]{get_ontology_levels} for more details.}

\item{add_info_content}{Add information content column for each phenotype.}

\item{add_description}{Whether to get the phenotype descriptions as
well (slower).}

\item{add_disease_data}{Add all disease metadata columns.
This will expand the data using \code{allow.cartesian=TRUE}.}

\item{add_ndiseases}{Add the number of diseases per phenotype.}

\item{add_onsets}{Add age of onset columns using
\link[HPOExplorer]{add_onset}.}

\item{add_deaths}{Add age of death columns using
\link[HPOExplorer]{add_death}.}

\item{add_pheno_frequencies}{Add the frequency of each phenotype in
each disease.}

\item{add_tiers}{Add severity Tiers column using
\link[HPOExplorer]{add_tier}.}

\item{add_severities}{Add severity column using
\link[HPOExplorer]{add_severity}.}

\item{add_disease_definitions}{Add disease definitions.}

\item{include_mondo}{Add \href{https://mondo.monarchinitiative.org/}{MONDO}
IDs, names, and definitions to each disease.}

\item{add_hoverboxes}{Add hoverdata with
\link[KGExplorer]{add_hoverboxes}.}

\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[KGExplorer]{add_hoverboxes}.}

\item{verbose}{Print messages.}

\item{terms}{A subset of HPO IDs to assign Tiers to.}

\item{keyword_sets}{A named list of regex queries to be used when searching
for phenotypes that have matching metadata.}

\item{check_names}{When regexsearching for matching terms,
check the phenotype names.}

\item{check_definitions}{When regex searching for matching terms,
check the phenotype definition.}

\item{check_diseases}{When regex searching for matching terms,
check the names of diseases associated with the phenotype.}

\item{search_ancestors}{Inherit Tiers of ancestors.}

\item{search_descendants}{Inherit Tiers of descendants.}

\item{as_datatable}{Return the results as a \link[data.table]{data.table}.}
}
\value{
R object.

A \link[igraph]{igraph} object.

A \link[ggnetwork]{ggnetwork} object.

A named list containing the graph data and the plot.

The HPO in dataframe format.

Tier assignments for each term in \code{terms}.
Will be returned as either a named vector or a \link[data.table]{data.table}.
}
\description{
Functions to make complex R objects.
}
\section{Functions}{
\itemize{
\item \code{make_igraph_object()}: make_
Make an \link[igraph]{igraph} object

This uses the network package to coerce the adjacency matrix into a
network object. It also adds the fold change, label,
and relative ontology level parameters to each node in the network.

It expects there to be a column of HPO IDs in the phenos dataframe called
hpo_id.

\item \code{make_network_object()}: make_
Make a \link[ggnetwork]{ggnetwork} object

This uses the network package to coerce the adjacency matrix into a
network object. It also adds the fold change, label,
and relative ontology level parameters to each node in the network.

It expects there to be a column of HPO IDs in the phenos dataframe called
hpo_id.

\item \code{make_network_plot()}: make_
Generate network plot

This creates a network plot.
It makes it possible to hover box that includes your results related to
each phenotype and/or a description of the phenotypes.

\item \code{make_phenos_dataframe()}: make_
Make phenotypes dataframe

Make a dataframe from a subset of the Human Phenotype Ontology.

\item \code{make_tiers()}: make_
Make severity Tiers (auto)

Automatically add severity Tier for each HPO ID, in accordance with the
rating system provided by
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4262393/}{
Lazarin et al (2014)}.
In order of increasing severity:
\itemize{
\item{Tier 4 }{Reduced fertility}
\item{Tier 3 }{Sensory impairment: vision,
Immunodeficiency/cancer,
Sensory impairment: hearing,
Sensory impairment: touch, other (including pain),
Mental illness,
Dysmorphic features}
\item{Tier 2 }{Shortened life span: premature adulthood,
Impaired mobility,
Internal physical malformation}
\item{Tier 1 }{Shortened life span: infancy,
Shortened life span: childhood/adolescence,
Intellectual disability}
}

}}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
g <- make_igraph_object(phenos = phenos)
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenoNet <- make_network_object(phenos = phenos,
                                colour_var = "ontLvl_geneCount_ratio")
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
plt <- make_network_plot(phenos = phenos)
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
terms <- get_hpo()@terms[seq(100)]
tiers <- make_tiers(terms = terms)
}
\concept{make_}
