% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hoverbox.R
\name{make_hoverbox}
\alias{make_hoverbox}
\title{Make Hoverbox}
\usage{
make_hoverbox(
  phenotype,
  phenos_dataframe,
  columns = c("HPO_Id", "description"),
  labels = c("ID", "Description")
)
}
\arguments{
\item{phenotype}{The phenotype you are making a hoverbox for}

\item{phenos_dataframe}{The dataframe of phenotypes and parameters to be included}

\item{columns}{chr vector of column names from the phenos_dataframe}

\item{labels}{chr vector of labels for the selected columns}
}
\value{
A nicely formatted string with newlines etc, to be used as a hoverbox
}
\description{
A hoverbox is a box of text that shows up when the cursor hovers over something.
These can be useful when making interactive network plots of the HPO phenotypes
because we can include a hoverbox that gives information and data associated
with each phenotype.
}
\details{
This function expects a dataframe of with a "Phenotype" column that has the
name of each phenotype. It must then include columns for all of the parameters
you wish to include in the hoverbox.
}
\examples{
library(ontologyIndex)
data(hpo)
phenotype_to_genes <- load_phenotype_to_genes()
Neuro_delay_ID <- get_hpo_termID_direct(hpo = hpo,
                                        phenotype = "Neurodevelopmental delay")
Neuro_delay_descendants <- phenotype_to_genes[
    phenotype_to_genes$ID \%in\% get_descendants(hpo,Neuro_delay_ID),]

phenos = data.frame()
for (p in unique(Neuro_delay_descendants$Phenotype)) {
    id <- get_hpo_termID(phenotype = p,
                         phenotype_to_genes = phenotype_to_genes)
    ontLvl_geneCount_ratio <- (get_ont_level(hpo = hpo,
                                             term_ids = p) + 1)/length(get_gene_list(p,phenotype_to_genes))
    description <- get_term_definition(ontologyId = id,
                                       line_length = 10)
    phenos <- rbind(phenos,
                    data.frame("Phenotype"=p,
                               "HPO_Id"=id,
                               "ontLvl_geneCount_ratio"=ontLvl_geneCount_ratio,
                               "description"=description))
}

hoverBox = c()
for (p in unique(phenos$Phenotype)) {
    hoverBox <- append(hoverBox,
                       make_hoverbox(p,
                                     phenos,
                                     columns = c("HPO_Id",
                                                 "ontLvl_geneCount_ratio",
                                                 "description"),
                                     labels = c("ID",
                                                "Ont.Lvl/n.genes",
                                                "Description")))
}
phenos$hover<- hoverBox


}
