% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_igraph_object.R
\name{make_igraph_object}
\alias{make_igraph_object}
\title{Make an \link[igraph]{igraph} object}
\usage{
make_igraph_object(
  phenos,
  hpo = get_hpo(),
  adjacency = adjacency_matrix(terms = phenos$hpo_id, hpo = hpo),
  colour_var = "fold_change",
  add_ont_lvl_absolute = FALSE,
  cols = list_columns(extra_cols = c(colour_var, grep("_count$|_values$", names(phenos),
    value = TRUE))),
  layout = "fruchtermanreingold",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{An adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{colour_var}{The column from phenos that you wish
to map to node colour.}

\item{add_ont_lvl_absolute}{Add the absolute ontology level of each HPO term.
See \link[HPOExplorer]{get_ont_lvls} for more details.}

\item{cols}{Columns to add to metadata of \link[ggnetwork]{ggnetwork} object.}

\item{layout}{a network layout supplied by \code{\link[sna]{gplot.layout}},
such as \code{"fruchtermanreingold"} (the default), or a two-column matrix
with as many rows as there are nodes in the network, in which case the
matrix is used as nodes coordinates.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[ggnetwork:ggnetwork]{ggnetwork::ggnetwork}}
  \describe{
    \item{\code{x}}{an object of class \code{\link[network]{network}} or
\code{\link[igraph:igraph-package]{igraph}}, or any object that can be
coerced to that class, such as an adjacency or incidence matrix, or an
edge list: see \code{\link[network]{edgeset.constructors}} and
\code{\link[network]{network}} for details.}
  }}
}
\value{
A \link[igraph]{igraph} object.
}
\description{
This uses the network package to coerce the adjacency matrix into a
network object. It also adds the fold change, label,
and relative ontology level parameters to each node in the network.
}
\details{
It expects there to be a column of HPO IDs in the phenos dataframe called
hpo_id.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
g <- make_igraph_object(phenos = phenos)
}
