% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phenos_dataframe.R
\name{make_phenos_dataframe}
\alias{make_phenos_dataframe}
\title{Make phenotypes dataframe}
\usage{
make_phenos_dataframe(
  hpo,
  phenotype_to_genes = load_phenotype_to_genes(),
  ancestor = NULL,
  add_description = TRUE,
  line_length = 10,
  verbose = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{ancestor}{The ancestor to get all descendants of. If \code{NULL},
returns the entirely ontology.}

\item{add_description}{Whether to get the phenotype descriptions as
well (slower).}

\item{line_length}{The number of desired words per line \<int\>}

\item{verbose}{Print messages.}
}
\value{
The HPO in dataframe format.
}
\description{
Make a dataframe from a subset of the Human Phenotype Ontology.
}
\examples{
library(ontologyIndex)
data(hpo)
pheno <- make_phenos_dataframe(hpo = hpo,
                               ancestor = "Neurodevelopmental delay",
                               add_description = FALSE)
}
