% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/list_deaths.R, R/list_onsets.R
\name{map_}
\alias{map_}
\alias{list_deaths}
\alias{list_onsets}
\title{Map functions}
\usage{
list_deaths(
  hpo = get_hpo(),
  exclude = FALSE,
  to = c("name", "id"),
  include_na = TRUE
)

list_onsets(
  hpo = get_hpo(),
  exclude = FALSE,
  to = c("name", "id"),
  include_na = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{exclude}{HPO phenotype names to exclude.}

\item{to}{Character vector of database(s) to map IDs to.
When not \code{"mondo"}, can supply multiple alternative databases to map to
 (e.g. \code{c("OMIM","Orphanet","DECIPHER")}).}

\item{include_na}{Include NA values for onset.}

\item{verbose}{Print messages.}
}
\value{
Mapped data.

Named list of HPO IDs.

Named list of HPO IDs.
}
\description{
Functions to map IDs.
}
\section{Functions}{
\itemize{
\item \code{list_deaths()}: map_
List age of death HPO terms

List age of death phenotypes in the HPO.

\item \code{list_onsets()}: map_
List age of onset HPO terms

List age of onset phenotypes in the HPO.

}}
\examples{
deaths <- list_deaths()
onsets <- list_onsets()
}
\concept{map_}
