% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_disease.R
\name{map_disease}
\alias{map_disease}
\title{Map disease}
\source{
\href{https://ontology.jax.org/api/network/docs}{HPO API docs}

\href{https://github.com/obophenotype/human-phenotype-ontology/issues/10232}{HPO GitHub Issue}
}
\usage{
map_disease(
  dat,
  id_col = "disease_id",
  fields = c("disease", "categories", "genes")[1],
  use_api = FALSE,
  return_dat = FALSE,
  workers = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE
)
}
\arguments{
\item{dat}{A data.table with a column of disease IDs.}

\item{id_col}{The name of the column with the disease IDs.}

\item{fields}{The fields to extract from the API response.}

\item{use_api}{Use the API to get the disease names and descriptions.
Otherwise, use a cached data.table (\code{disease_map}).}

\item{return_dat}{Return the data.table with the mapped fields.}

\item{workers}{Number (>1) or proportion (<1) of worker cores to use.}

\item{all.x}{logical; if \code{TRUE}, rows from \code{x} which have no matching row
in \code{y} are included. These rows will have 'NA's in the columns that are usually
filled with values from \code{y}. The default is \code{FALSE} so that only rows with
data from both \code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}
}
\description{
Map disease IDs (e.g. "OMIM:101200") to names (e.g. "Apert syndrome")
}
\examples{
dat <- HPOExplorer::load_phenotype_to_genes()
out <- map_disease(dat = dat, workers=1)
}
