% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_phenotypes.R
\name{map_phenotypes}
\alias{map_phenotypes}
\title{Harmonise phenotypes}
\usage{
map_phenotypes(
  terms,
  hpo = get_hpo(),
  to = c("name", "id"),
  keep_order = TRUE,
  ignore_case = TRUE,
  ignore_char = eval(formals(KGExplorer::map_ontology_terms)$ignore_char),
  invert = FALSE
)
}
\arguments{
\item{terms}{A subset of HPO IDs to include in the final dataset and plots
(e.g. c("HP:0001508","HP:0001507")).}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{to}{Character vector of database(s) to map IDs to.
When not \code{"mondo"}, can supply multiple alternative databases to map to
 (e.g. \code{c("OMIM","Orphanet","DECIPHER")}).}

\item{keep_order}{Return a named list of the same length and order
as \code{terms}.
If \code{FALSE}, return a named list of only the unique \code{terms},
sometimes in a different order.}

\item{ignore_case}{Ignore case when mapping terms.}

\item{ignore_char}{A character vector of characters to ignore when 
mapping terms.}

\item{invert}{Invert the keys/values of the dictionary,
such that the key becomes the values (and vice versa).}
}
\value{
Character vector
}
\description{
Harmonise a mixed vector of phenotype names (e.g. "Focal motor seizure")
and HPO IDs (e.g. c("HP:0000002","HP:0000003")).
}
\examples{
terms <- c("Focal motor seizure",
            "Focal MotoR SEIzure",
            "HP:0000002","HP:0000003")
#### As phenotype names ####
term_names <- map_phenotypes(terms=terms)
#### As HPO IDs ####
term_ids <- map_phenotypes(terms=terms, to="id")
}
