% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_upheno.R
\name{map_upheno}
\alias{map_upheno}
\title{Map phenotypes across uPheno}
\source{
https://data.monarchinitiative.org/upheno2/current/qc/index

https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html
}
\usage{
map_upheno(
  pheno_map_method = c("upheno", "monarch"),
  gene_map_method = c("monarch", "orthogene"),
  subset_db1 = c("HP"),
  terms = NULL,
  fill_scores = NULL,
  show_plot = TRUE,
  force_new = FALSE,
  save_dir = tools::R_user_dir(package = "HPOExplorer", which = "cache")
)
}
\arguments{
\item{pheno_map_method}{Method to use for mapping phenotypes across ontologies.
\itemize{
\item{"upheno"}{Use uPheno's phenotype-to-phenotype mappings.
Contains fewer ontologies but with greater coverage of phenotypes.}
\item{"monarch"}{Use Monarch's phenotype-to-phenotype mappings.
Contains more ontologies but with less coverage of phenotypes.
}
}}

\item{gene_map_method}{Method to use for mapping genes across species.
\itemize{
\item{"monarch"}{Use Monarch's gene-to-gene mappings.}
\item{"orthogene"}{Use \link[orthogene]{convert_orthologs} to generate
gene-to-gene mappings.}
}}

\item{subset_db1}{Subset of ontologies to include in the plot.}

\item{terms}{A subset of HPO IDs to include in the final dataset and plots
(e.g. c("HP:0001508","HP:0001507")).}

\item{fill_scores}{Fill missing scores in the "equivalence_score" and
"subclass_score" columns with this value. These columns represent the
quality of mapping between two phenotypes on a scale from 0-1.}

\item{show_plot}{Show the plot.}

\item{force_new}{Force new data to be downloaded and processed.}

\item{save_dir}{Directory to save cached data.}
}
\value{
A list containing the data and plot.
}
\description{
Map phenotypes across species within the Unified Phenotype Ontology (uPheno).
First, gathers phenotype-phenotype mappings across ontologies.
Next, gathers all phenotype-gene associations for each ontology,
converts all genes to human HGNC orthologs, and computes the number of
overlapping orthologs between all pairs mapped phenotypes.
Finally, plots the results as the proportion of intersecting genes between
all pairs of phenotypes.
}
\examples{
res <- map_upheno()
}
