% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_upheno_data.R
\name{map_upheno_data}
\alias{map_upheno_data}
\title{Map uPheno data}
\usage{
map_upheno_data(
  pheno_map_method = c("upheno", "monarch"),
  gene_map_method = c("monarch", "orthogene"),
  keep_nogenes = FALSE,
  fill_scores = NULL,
  terms = NULL,
  save_dir = tools::R_user_dir(package = "HPOExplorer", which = "cache"),
  force_new = FALSE
)
}
\arguments{
\item{pheno_map_method}{Method to use for mapping phenotypes across ontologies.
\itemize{
\item{"upheno"}{Use uPheno's phenotype-to-phenotype mappings.
Contains fewer ontologies but with greater coverage of phenotypes.}
\item{"monarch"}{Use Monarch's phenotype-to-phenotype mappings.
Contains more ontologies but with less coverage of phenotypes.
}
}}

\item{gene_map_method}{Method to use for mapping genes across species.
\itemize{
\item{"monarch"}{Use Monarch's gene-to-gene mappings.}
\item{"orthogene"}{Use \link[orthogene]{convert_orthologs} to generate
gene-to-gene mappings.}
}}

\item{keep_nogenes}{Logical indicating whether to keep mappings that do not
have any orthologous genes.}

\item{fill_scores}{Fill missing scores in the "equivalence_score" and
"subclass_score" columns with this value. These columns represent the
quality of mapping between two phenotypes on a scale from 0-1.}

\item{terms}{A subset of HPO IDs to include in the final dataset and plots
(e.g. c("HP:0001508","HP:0001507")).}

\item{save_dir}{Directory to save cached data.}

\item{force_new}{Force new data to be downloaded and processed.}
}
\value{
A data.table containing the mapped data.
}
\description{
Get uPheno cross-species mapping data by:
\itemize{
\item{Downloading cross-species phenotype-phenotype mappings.}
\item{Downloading within-spceies phenotype-gene mappings,
and converting these genes to human orthologs.}
\item{Merging the phenotype-phenotype and phenotype-gene mappings.}
\item{Filtering out any mappings that do not have a human ortholog within
each respective phenotype.}
\item{Calculating the proportion of orthologous genes overlapping across
the species-specific phenotype-gene mappings.}
\item{Iterating the above steps using multiple methods
(\code{pheno_map_method}) and concatenating the results together.}
}
}
\examples{
pheno_map_genes_match <- map_upheno_data()
}
