% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_upheno_plots.R
\name{map_upheno_plot}
\alias{map_upheno_plot}
\title{Map uPheno plot}
\usage{
map_upheno_plot(
  pheno_map_genes_match = NULL,
  subset_db1 = "HP",
  types = c("rainplot", "scatterplot", "heatmap")
)
}
\arguments{
\item{pheno_map_genes_match}{uPheno cross-species mapping data generated by
\link[HPOExplorer]{map_upheno_data}.}

\item{subset_db1}{A character vector of reference ontologies to subset
\code{pheno_map_genes_match} by.}

\item{types}{A character vector of plot types to generate.}
}
\value{
A named list of \pkg{ggplot2} objects.
}
\description{
Generate multiple kinds of plots summarising mappings of phenotypes and genes
between the HPO and various non-human ontologies within uPheno.
}
\examples{
pheno_map_genes_match <- map_upheno_data()
upheno_plots <- map_upheno_plot(
  pheno_map_genes_match = pheno_map_genes_match)
}
