% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_3d.R
\name{network_3d}
\alias{network_3d}
\title{3D network}
\source{
\href{https://www.r-bloggers.com/2013/09/network-visualization-part-4-3d-networks/}{
R bloggers}

\href{https://community.plotly.com/t/is-it-possible-to-connect-scatters-in-3d-scatter-plot/}{
Plotly: Connect points in 3D plot}

\href{https://plotly.com/r/line-charts/}{
Plotly: Connect points in 2D plot}

\href{https://plotly.com/r/3d-line-plots/}{
Plotly: 3D lines plots}
}
\usage{
network_3d(
  g,
  layout_func = igraph::layout.fruchterman.reingold,
  extend_kde = 1.2,
  add_labels = FALSE
)
}
\arguments{
\item{g}{\link[igraph]{igraph} object
generated by \link[HPOExplorer]{make_igraph_object}.}

\item{layout_func}{Layout function for the graph.}

\item{extend_kde}{Extend the area that the
kernel density estimation (KDE) plot
(i.e. the "mountains" underneath the points) covers.}

\item{add_labels}{Add phenotype name labels to each point.}
}
\value{
plotly object.
}
\description{
Plot a subset of the HPO as a 3D network.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
g <- make_igraph_object(phenos = phenos)
}
