% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newlines_to_definition.R
\name{newlines_to_definition}
\alias{newlines_to_definition}
\title{Add new lines to disease description}
\usage{
newlines_to_definition(definition, line_length = 10)
}
\arguments{
\item{definition}{A disease description string}

\item{line_length}{A integer representing the desired words per line.}
}
\value{
The disease description with newline symbols added every nth word.
}
\description{
Adds new lines to the description so that hover boxes don't get too wide.
}
\examples{
description <- "A dialeptic seizure is a type of seizure characterised
predominantly by reduced responsiveness or awareness and with subsequent at
least partial amnesia of the event."
newlines_to_definition(description, 10)
}
