% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_branch_plot.R
\name{per_branch_plot}
\alias{per_branch_plot}
\title{Plot N phenotypes per branch}
\usage{
per_branch_plot(
  highlighted_branches,
  ancestor,
  hpo = get_hpo(),
  background_branches = simona::dag_children(hpo, term = map_phenotypes(terms = ancestor,
    hpo = hpo, to = "id")),
  phenotype_to_genes = load_phenotype_to_genes()
)
}
\arguments{
\item{highlighted_branches}{Names of branches you want
to highlight \<vector\<string\>\>}

\item{ancestor}{The ancestor to get all descendants of. If \code{NULL},
returns the entirely ontology.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{background_branches}{HPO Ids of all branches to be shown
(including highlighted) \<vector\<string\>\>}

\item{phenotype_to_genes}{Output of
\link{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}
}
\value{
ggplot object
}
\description{
This plots the number of phenotypes per HPO branch and colours selected
branches to highlight them. This is not using any EWCE results, it is simply
using the raw HPO data to get an overview of how many phenotypes there are.
}
\details{
Default value for the background_branches argument is the child terms of
phenotypic abnormality. Essentially this gives the main branches of the HPO.

Note that the highlighted branches must be also present in the background
branches.
}
\examples{
plt <-  per_branch_plot(
    highlighted_branches = "Abnormality of nervous system physiology",
    ancestor = "Abnormality of the nervous system")
}
