% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arrow.R
\name{plot_arrow}
\alias{plot_arrow}
\title{Plot arrow}
\usage{
plot_arrow(
  x = 1,
  xend = 1,
  y = 0,
  yend = 1,
  labels_x = c(x, xend),
  labels_y = c(yend * 0.8, yend * 0.2),
  labels = c("Broad\\nphenotypes", "Specific\\nphenotypes"),
  labels_alpha = 0.8,
  labels_size = 4,
  arrrow_size = 2,
  arrow_color = "grey50"
)
}
\arguments{
\item{x}{x-coordinate of the arrow.}

\item{xend}{x-coordinate of the arrow end.}

\item{y}{y-coordinate of the arrow.}

\item{yend}{y-coordinate of the arrow end.}

\item{labels_x}{x-coordinates of the labels.}

\item{labels_y}{y-coordinates of the labels.}

\item{labels}{Labels for each end of the arrow.}

\item{labels_alpha}{Transparency of the labels.}

\item{labels_size}{Size of the labels.}

\item{arrrow_size}{Size of the arrow.}

\item{arrow_color}{Colour of the arrow.}
}
\value{
ggplot2 object.
}
\description{
Plot an error indicating the direction of phenotype specificity.
}
\examples{
plot_arrow()
}
