% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hpo.R
\name{search_hpo}
\alias{search_hpo}
\title{Search the Human Phenotype Ontology}
\usage{
search_hpo(
  hpo = get_hpo(),
  queries = list(intellectual_disability = c("intellectual disability"),
    impaired_mobility = c("Abnormal central motor function", "Abnormality of movement"),
    physical_malformations = c("malformation"), blindness = c("^blindness"),
    sensory_impairments = c("Abnormality of vision", "Abnormality of the sense of smell",
    "Abnormality of taste sensation", "Somatic sensory dysfunction",
    "Hearing abnormality"), immunodeficiency = c("Immunodeficiency"), cancer =
    c("Neoplasm", "Cancer"), reduced_fertility = c("Decreased fertility")),
  include_descendants = TRUE,
  include_ancestors = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{queries}{A named list of character vectors.
Each vector includes a phenoptype name (or a substring)
to be searched for in the HPO.}

\item{include_descendants}{Include all descendants of query hits.}

\item{include_ancestors}{Include all ancestors of query hits.}

\item{verbose}{Print messages.}
}
\value{
Named list of HPO IDs.
}
\description{
Serarch for phenotypes in the HPO that match substring queries,
or are the descendants of those substring-matched phenotypes.
}
\examples{
query_hits <- search_hpo()
}
