% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_descendants.R
\name{subset_descendants}
\alias{subset_descendants}
\title{Subset descendants}
\usage{
subset_descendants(
  phenos,
  ancestor = NULL,
  hpo = get_hpo(),
  ignore_case = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{ancestor}{Phenotype names (or HPO ID) of an ancestor in the
\href{https://hpo.jax.org/}{Human Phenotype Ontology}.
Only phenotypes that are descendants of this ancestor will be kept.
Set to \code{NULL} (default) to skip this filtering step.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{ignore_case}{Ignore casing when searching for matches.}

\item{verbose}{Print messages.}
}
\value{
data.table of phenotypes, with additional columns:
"ancestor", "ancestor_id"
}
\description{
Subset a \code{phenos} datatable to only
descendants of an ancestor HPO ID term.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- subset_descendants(phenos = phenos,
                              ancestor = "Motor delay")
}
