% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_col.R
\name{unlist_col}
\alias{unlist_col}
\title{Unlist column}
\usage{
unlist_col(dat, col)
}
\arguments{
\item{dat}{\link[data.table]{data.table}}

\item{col}{Column name.}
}
\value{
\link[data.table]{data.table}
}
\description{
Unlist a column in a \link[data.table]{data.table}.
}
\keyword{internal}
