% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_disease.R
\name{add_disease}
\alias{add_disease}
\title{Add diseases}
\usage{
add_disease(
  phenos,
  phenotype_to_genes = load_phenotype_to_genes(),
  hpo = get_hpo(),
  add_descriptions = FALSE,
  all.x = TRUE,
  use_api = FALSE,
  workers = NULL,
  allow.cartesian = FALSE
)
}
\arguments{
\item{phenos}{A data.table containing HPO IDs and other metadata.}

\item{phenotype_to_genes}{Output of
\link{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{add_descriptions}{Add disease names and descriptions.}

\item{all.x}{logical; if \code{TRUE}, rows from \code{x} which have no matching row
in \code{y} are included. These rows will have 'NA's in the columns that are usually
filled with values from \code{y}. The default is \code{FALSE} so that only rows with
data from both \code{x} and \code{y} are included in the output.}

\item{use_api}{Get definitions from the HPO API,
as opposed to a static local dataset.}

\item{workers}{Number (>1) or proportion (<1) of worker cores to use.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}
}
\description{
Annotate each HPO term with diseases that they are associated with.
}
\examples{
phenos <- example_phenos()
phenos2 <- add_disease(phenos = phenos)
}
