% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mondo.R
\name{add_mondo}
\alias{add_mondo}
\title{Add age of death}
\usage{
add_mondo(
  phenos,
  add_definitions = TRUE,
  all.x = TRUE,
  allow.cartesian = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{add_definitions}{Add disease definitions using
\link[HPOExplorer]{add_disease_definition}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{verbose}{Print messages.}
}
\description{
Add metadata from MONDO, including:
\itemize{
\item{MONDO_ID: }{MONDO term ID.}
\item{MONDO_name: }{MONDO term name.}
\item{MONDO_definition: }{MONDO term definition.}
}
}
\examples{
phenos <- load_phenotype_to_genes("phenotype.hpoa")
phenos2 <- add_mondo(phenos = phenos)
}
