% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_onset.R
\name{add_onset}
\alias{add_onset}
\title{Add age of onset}
\usage{
add_onset(
  phenos,
  keep_onsets = NULL,
  agg_by = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{keep_onsets}{The age of onset associated with each HPO ID to keep.
If >1 age of onset is associated with the term,
only the earliest age is considered.
See \link[HPOExplorer]{add_onset} for details.}

\item{agg_by}{Column to aggregate age of onset metadata by.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra columns:
\itemize{
\item{"onset": }{onset HPO IDs of disease phenotypes associated
with the target hpo_id phenotype.}
\item{"onset_names": }{onset HPO names of disease phenotypes associated
with the target hpo_id phenotype.}
\item{"onset_counts": }{The number of times each term in
"onset_names" appears across associated disease phenotypes.}
\item{"onset_score_mean": }{Mean onset score.}
\item{"onset_score_min": }{Minimum onset score.}
\item{"onset_top": }{The most common onset term.}
\item{"onset_earliest": }{The earliest age of onset.}
\item{"onset_latest": }{The latest age of onset.}
}
}
\description{
Add age of onset for each HPO ID.
onset IDs and assigned "onset_score" values:
\itemize{
\item{HP:0011461 }{"Fetal onset" (onset_score=1)}
\item{HP:0030674 }{"Antenatal onset" (onset_score=2)}
\item{HP:0003577 }{"Congenital onset" (onset_score=3)}
\item{HP:0003623 }{"Neonatal onset" (onset_score=4)}
\item{HP:0003593 }{"Infantile onset" (onset_score=5)}
\item{HP:0011463 }{"Childhood onset" (onset_score=6)}
\item{HP:0003621 }{"Juvenile onset" (onset_score=7)}
\item{HP:0011462 }{"Young adult onset" (onset_score=8)}
\item{HP:0003581 }{"Adult onset" (onset_score=9)}
\item{HP:0003596 }{"Middle age onset" (onset_score=10)}
\item{HP:0003584 }{"Late onset" (onset_score=11)}
}
}
\examples{
phenos <- example_phenos()
phenos2 <- add_onset(phenos = phenos)
}
