% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_onset.R
\name{add_onset}
\alias{add_onset}
\title{Add age of onset}
\usage{
add_onset(phenos, all.x = TRUE, allow.cartesian = TRUE, verbose = TRUE)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add age of onset for each HPO ID.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_onset(phenos = phenos)
}
