% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ont_lvl.R
\name{add_ont_lvl}
\alias{add_ont_lvl}
\title{Add ontology level}
\usage{
add_ont_lvl(
  phenos,
  hpo = get_hpo(),
  adjacency = NULL,
  absolute = TRUE,
  exclude_top_lvl = TRUE,
  keep_ont_levels = NULL,
  reverse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{An adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{absolute}{Make the levels absolute in the sense that they consider
the entire HPO ontology (\code{TRUE}).
Otherwise, levels will be relative to only the HPO terms that are in
the provided subset of \code{terms} AND are directly adjacent (connected)
to a given cluster of terms (\code{FALSE}).}

\item{exclude_top_lvl}{Exclude the top level term of the HPO
(i.e. "All" (HP:0000001)) when computing \emph{absolute} levels.
This argument is not used when computing \emph{relative} levels.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link[HPOExplorer]{add_ont_lvl} for details.}

\item{reverse}{If \code{TRUE}, ontology
level numbers with be revered such that the level of the parent terms
are larger than the child terms.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add the relative ontology level for each HPO ID.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_ont_lvl(phenos = phenos)
}
