% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tier.R
\name{add_tier}
\alias{add_tier}
\title{Add severity Tiers}
\usage{
add_tier(
  phenos,
  all.x = TRUE,
  include_disease_characteristics = TRUE,
  auto_assign = TRUE,
  hpo = get_hpo(),
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{include_disease_characteristics}{Include \code{phenotypes} that
are also high-level \code{include_disease_characteristics}.}

\item{auto_assign}{Automatically assing HPO IDs to Tiers by conducting
regex searches for keywords that appear in the term name,
or the names of its descendants or ancestors.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{verbose}{Print messages.}
}
\value{
phenos data.table with extra column
}
\description{
Add severity Tier for each HPO ID, in accordance with the
rating system provided by
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4262393/}{
Lazarin et al (2014)}.
In order of increasing severity:
\itemize{
\item{Tier 4 }{Reduced fertility}
\item{Tier 3 }{Sensory impairment: vision,
Immunodeficiency/cancer,
Sensory impairment: hearing,
Sensory impairment: touch, other (including pain),
Mental illness,
Dysmorphic features}
\item{Tier 2 }{Shortened life span: premature adulthood,
Impaired mobility,
Internal physical malformation}
\item{Tier 1 }{Shortened life span: infancy,
Shortened life span: childhood/adolescence,
Intellectual disability}
}
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- add_tier(phenos = phenos)
}
