% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Adjacency matrix}
\usage{
adjacency_matrix(pheno_ids, hpo, as_dataframe = FALSE)
}
\arguments{
\item{pheno_ids}{a character vector of HPO Ids}

\item{hpo}{ontology object (available in ontologyIndex package)}

\item{as_dataframe}{can return matrix or df <bool>}
}
\value{
adjacency matrix
}
\description{
Create adjacency matrix of HPO child-parent relationships
}
\details{
It may be possible to use a hash table for ggnetwork, which may be more
efficent than the matrix in a shiny app ?
}
\examples{
\dontrun{
pheno_ids <- c("HP:000001","HP:000002")
adjacency_matrix(pheno_ids,hpo,as_dataframe=FALSE)
}
}
