% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Adjacency matrix}
\usage{
adjacency_matrix(pheno_ids, hpo)
}
\arguments{
\item{pheno_ids}{a character vector of HPO Ids}

\item{hpo}{ontology object (available in ontologyIndex package)}
}
\value{
adjacency matrix
}
\description{
Create adjacency matrix of HPO child-parent relationships
}
\details{
This is needed for plots created using ggnetwork as it can coerce an
adjacency matrix into a directed graph object, and assigns
each datapoint an x,y coordinate. Maybe shouldn't use a for loop for this.
It also may be possible to use a hash table for ggnetwork, which may be more
efficent for the web app
}
\examples{
\dontrun{
adjacency_matrix(c("HP:000001","HP:000002"),hpo)
}
}
