% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Adjacency matrix}
\usage{
adjacency_matrix(pheno_ids, hpo = get_hpo(), as_dataframe = FALSE)
}
\arguments{
\item{pheno_ids}{Character vector of HPO IDs.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{as_dataframe}{Can return matrix or dataframe \<bool\>}
}
\value{
adjacency matrix
}
\description{
Create adjacency matrix of HPO child-parent relationships.
}
\examples{
adjacency <- adjacency_matrix(pheno_ids = c("HP:000001", "HP:000002"))
}
