% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/filter_descendants.R
\name{filter_}
\alias{filter_}
\alias{filter_descendants}
\title{Filter functions}
\usage{
filter_descendants(phenos, ancestor = NULL, hpo = get_hpo())
}
\arguments{
\item{phenos}{A data.table containing HPO IDs and other metadata.}

\item{ancestor}{Phenotype names (or HPO ID) of an ancestor in the
\href{https://hpo.jax.org/}{Human Phenotype Ontology}.
Only phenotypes that are descendants of this ancestor will be kept.
Set to \code{NULL} (default) to skip this filtering step.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}
}
\value{
Filtered data.

data.table of phenotypes, with additional columns:
"ancestor", "ancestor_id"
}
\description{
Functions to filter data.table objects.
}
\section{Functions}{
\itemize{
\item \code{filter_descendants()}: filter_
Filter descendants

Subset a \code{phenos} data.table to only
descendants of an ancestor HPO ID term.

}}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenos2 <- filter_descendants(phenos = phenos,
                              ancestor = "Motor delay")
}
\concept{filter_}
