% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_hpo_ids.R
\name{fix_hpo_ids}
\alias{fix_hpo_ids}
\title{Fix missing HPO IDs}
\usage{
fix_hpo_ids(dt, phenotype_to_genes = load_phenotype_to_genes())
}
\arguments{
\item{dt}{\link[data.table]{data.table}
containing a column named "Phenotype".}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}
}
\value{
data.table with the column "HPO_ID"
}
\description{
Fix missing HPO IDs by inferring their identity from the \code{Phenotype}
column using several methods.
}
\keyword{internal}
