% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/get_hpo.R, R/get_hpo_id_direct.R
\name{get_}
\alias{get_}
\alias{get_hpo}
\alias{get_hpo_id_direct}
\title{Get functions}
\usage{
get_hpo(
  lvl,
  add_ancestors = 2,
  force_new = FALSE,
  terms = NULL,
  method = "github",
  save_dir = KGExplorer::cache_dir(package = "HPOExplorer"),
  ...
)

get_hpo_id_direct(term, hpo = get_hpo())
}
\arguments{
\item{lvl}{How many levels deep into the ontology to get ancestors from.
For example:
\itemize{
\item{1: }{"All"}
\item{2: }{"Phenotypic abnormality"}
\item{3: }{"Abnormality of the nervous system"}
\item{4: }{"Abnormality of nervous system physiology"}
\item{5: }{"Neurodevelopmental abnormality" or "Behavioral abnormality"}
}}

\item{add_ancestors}{Add ancestors for each term.}

\item{force_new}{Force a new download or creation of a data resource.}

\item{terms}{A subset of HPO IDs to assign Tiers to.}

\item{method}{Method to construct plot with.}

\item{save_dir}{Directory to save a file to.}

\item{...}{
  Arguments passed on to \code{\link[KGExplorer:get_]{KGExplorer::get_ontology}}
  \describe{
    \item{\code{name}}{\itemize{
\item{<...>}{Any ontology name from \link[KGExplorer]{get_ols_options}}
\item{"mondo"}{
Import the \href{https://mondo.monarchinitiative.org/}{Mondo} ontology.
\href{https://github.com/monarch-initiative/mondo/release}{
GitHub.}}.
\item{"hpo"}{
Import the \href{https://hpo.jax.org/app/}{Human Phenotype Ontology}.
\href{https://github.com/obophenotype/human-phenotype-ontology/release}{
GitHub.}}
}}
    \item{\code{filetype}}{File type to search for.}
    \item{\code{add_metadata}}{Add metadata to the resulting ontology object.}
    \item{\code{add_n_edges}}{Add the number of edges (connections) for each term.}
    \item{\code{add_ontology_levels}}{Add the ontology level for each term.}
  }}

\item{term}{One or more ontology IDs.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}
}
\value{
Data.

\link[simona]{ontology_DAG} object.

The HPO ID(s) of phenotype(s)
}
\description{
Functions to get data objects and extract elements.
}
\section{Functions}{
\itemize{
\item \code{get_hpo()}: get_
Get Human Phenotype Ontology (HPO)

Updated version of Human Phenotype Ontology (HPO).
Created from the OBO files distributed by the HPO project's
\href{https://github.com/obophenotype/human-phenotype-ontology}{GitHub}.

By comparison, the \code{hpo} data from \pkg{ontologyIndex} is from 2016.
Note that the maximum ontology level depth in the 2016 version was 14,
whereas in the 2023 version the maximum ontology level depth is 16
(due to an expansion of the HPO).

\item \code{get_hpo_id_direct()}: get_
Get HPO term ID direct

Directly retrieves it from the HPO
ontology object.

}}
\examples{
hpo <- get_hpo()
term = "Phenotypic abnormality"
pheno_abnormality_id <- get_hpo_id_direct(term = term)
}
\concept{get_}
