% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get remote data}
\usage{
get_data(
  file,
  tag = "latest",
  repo = "neurogenomics/HPOExplorer",
  save_dir = tools::R_user_dir(package = "HPOExplorer", which = "cache"),
  add_version = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{save_dir}{Directory to save data to.}

\item{add_version}{Add the release version
to the returned object's \link[base]{attributes}}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
}
\value{
Path to downloaded file or the object itself (when ".rds" format).
}
\description{
Download remotely stored data via \link[piggyback]{pb_download}.
}
\keyword{internal}
