% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpo_id_direct.R
\name{get_hpo_id_direct}
\alias{get_hpo_id_direct}
\title{Get HPO term ID direct}
\usage{
get_hpo_id_direct(phenotype, hpo = get_hpo())
}
\arguments{
\item{phenotype}{One or more phenotype names in to get HPO IDs for.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}
}
\value{
The HPO ID(s) of phenotype(s)
}
\description{
Directly retrieves it from the HPO
ontology object (found in the \pkg{ontologyIndex} package \code{data(hpo)}).
}
\examples{
phenotype = "Phenotypic abnormality"
pheno_abnormality_id <- get_hpo_id_direct(phenotype = phenotype)
}
