% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mondo.R
\name{get_mondo}
\alias{get_mondo}
\title{Get Mondo Disease Ontology}
\source{
\code{
save_path <- file.path(tempdir(),"mondo.rds")
URL <- paste0(
"https://data.bioontology.org/ontologies/MONDO/submissions/52/download",
"?apikey=8b5b7825-538d-40e0-9e9e-5ab9274a9aeb"
)
tmp <- file.path(tempdir(),"mondo.obo")
utils::download.file(URL, tmp)
mondo <- ontologyIndex::get_OBO(file = tmp, extract_tags="everything")
saveRDS(mondo,save_path)
piggyback::pb_new_release(repo = "neurogenomics/HPOExplorer",
                          tag = "latest")
piggyback::pb_upload(file = save_path,
                     repo = "neurogenomics/HPOExplorer",
                     overwrite = TRUE,
                     tag = "latest")
}
}
\usage{
get_mondo(
  save_dir = tools::R_user_dir("HPOExplorer", which = "cache"),
  tag = "latest",
  overwrite = TRUE
)
}
\arguments{
\item{save_dir}{Directory to save data to.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
}
\value{
\link[ontologyIndex]{ontology_index}
}
\description{
Import \href{https://mondo.monarchinitiative.org/}{Mondo} as an R object.
}
\examples{
mondo <- get_mondo()
}
