% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_check.R
\name{gpt_annot_check}
\alias{gpt_annot_check}
\title{Check annotations from chatGPT}
\usage{
gpt_annot_check(path, query_hits = search_hpo(), verbose = TRUE)
}
\arguments{
\item{path}{Path to annotations file.}
}
\value{
Named list
}
\description{
Check chatGPT phenotype annotations using a several metrics.
}
\keyword{internal}
