% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_check.R
\name{gpt_annot_check}
\alias{gpt_annot_check}
\title{Check annotations from GPT}
\usage{
gpt_annot_check(
  annot = gpt_annot_read(),
  query_hits = search_hpo(),
  response_map = list(no = "never", yes = "always"),
  pos_values = c("sometimes", "often", "always"),
  neg_values = c("never", "rarely"),
  show_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{annot}{GPT-generated phenotype annotations.}

\item{query_hits}{A named list of HPO IDs for each query,
generated by \link[HPOExplorer]{search_hpo}.
These will be used as the ground truth when trying to identify
true positive annotations.}

\item{response_map}{A named list of responses to map onto.
Used for standardisation of responses.}

\item{pos_values}{Positive values.}

\item{neg_values}{Negative values.}

\item{show_plot}{Include a plot of the results.}

\item{verbose}{Print messages.}
}
\value{
Named list
}
\description{
Check GPT phenotype annotations using a several metrics.
}
\examples{
checks <- gpt_annot_check()
}
