% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_codify.R
\name{gpt_annot_codify}
\alias{gpt_annot_codify}
\title{Check annotations from GPT}
\usage{
gpt_annot_codify(
  annot = gpt_annot_read(),
  remove_duplicates = TRUE,
  code_dict = c(never = 0, rarely = 1, varies = 2, often = 3, always = 4),
  tiers_dict = list(intellectual_disability = 1, death = 1, impaired_mobility = 2,
    physical_malformations = 2, blindness = 3, sensory_impairments = 3, immunodeficiency
    = 3, cancer = 3, reduced_fertility = 4),
  keep_congenital_onset = head(names(code_dict), 4)
)
}
\arguments{
\item{annot}{GPT-generated phenotype annotations.}

\item{remove_duplicates}{Ensure only 1 row per phenotype.}

\item{code_dict}{Numerical encodings of annotation values.}

\item{tiers_dict}{Numerical encodings of annotation column.}

\item{keep_congenital_onset}{Which stages of congenital onset to keep.}
}
\value{
Named list
}
\description{
Check GPT phenotype annotations using a several metrics.
}
\examples{
coded <- gpt_annot_codify()
}
