% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_plot.R
\name{gpt_annot_plot}
\alias{gpt_annot_plot}
\title{Plot annotations from GPT}
\usage{
gpt_annot_plot(
  annot = gpt_annot_read(),
  keep_ont_levels = seq(3, 17),
  keep_descendants = "Phenotypic abnormality",
  top_n = 50,
  verbose = TRUE
)
}
\arguments{
\item{annot}{GPT-generated phenotype annotations.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link{add_ont_lvl} for details.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{top_n}{Top number of most severe phenotypes to plot in heatmap.}

\item{verbose}{Print messages.}
}
\value{
Named list of plots.
}
\description{
Plot annotations from GPT.
}
\examples{
plots <- gpt_annot_plot()
}
