% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_read.R
\name{gpt_annot_read}
\alias{gpt_annot_read}
\title{Read annotations from GPT}
\usage{
gpt_annot_read(
  save_path = file.path(KGExplorer::cache_dir(package = "HPOExplorer"),
    "gpt4_hpo_annotations.csv"),
  phenotype_to_genes = load_phenotype_to_genes(),
  force_new = FALSE,
  hpo = get_hpo(),
  include_nogenes = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{save_path}{Path to annotations CSV file.
If the file does not exist, the data will be downloaded from GitHub.}

\item{phenotype_to_genes}{Output of
\link{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{force_new}{If \code{TRUE}, the data will be downloaded from GitHub
even if it already exists locally.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{include_nogenes}{Include phenotypes with no associated genes.}

\item{verbose}{Print messages.}
}
\value{
data.table of phenotype annotations
}
\description{
Read in phenotype annotations generated by GPT and
do some initial preprocessing (e.g. adding HPO IDs).
}
\examples{
gpt_annot <- gpt_annot_read()
}
