% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_annot_read.R
\name{gpt_annot_read}
\alias{gpt_annot_read}
\title{Read annotations from GPT}
\source{
code {
piggyback::pb_upload(file = "~/Downloads/gpt_hpo_annotations.csv",
repo = "neurogenomics/HPOExplorer",
overwrite = TRUE,
tag = "latest")
}
}
\usage{
gpt_annot_read(path = NULL, verbose = TRUE)
}
\arguments{
\item{path}{Path to annotations CSV file.
If \code{NULL}, will pull data from GitHub Releases instead.}

\item{verbose}{Print messages.}
}
\value{
data.table of phenotype annotations
}
\description{
Read in phenotype annotations generated by GPT and
do some initial preprocessing (e.g. adding HPO IDs).
}
\examples{
annot <- gpt_annot_read()
}
