% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_phenotypes.R
\name{harmonise_phenotypes}
\alias{harmonise_phenotypes}
\title{Harmonise phenotypes}
\usage{
harmonise_phenotypes(
  phenotypes,
  hpo = get_hpo(),
  as_hpo_ids = FALSE,
  validate = TRUE,
  ignore_case = TRUE,
  keep_order = TRUE,
  invert = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenotypes}{A character vector of phenotype names and/or HPO IDs.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{as_hpo_ids}{Return all \code{phenotypes} as HPO IDs (\code{TRUE}),
or phenotype names (\code{FALSE}).}

\item{validate}{Remove any phenotype names or IDs that cannot be found in the
\code{hpo}.}

\item{ignore_case}{Ignore casing when searching for matches.}

\item{keep_order}{Return a named list of the same length and order
as \code{phenotypes}.
If \code{FALSE}, return a named list of only the unique \code{phenotypes},
sometimes in a different order.}

\item{invert}{Invert the keys/values of the dictionary,
such that the key becomes the values (and vice versa).}

\item{verbose}{Print messages.}
}
\value{
Character vector
}
\description{
Harmonise a mixed vector of phenotype names (e.g. "Focal motor seizure")
and HPO IDs (e.g. c("HP:0000002","HP:0000003")).
}
\examples{
phenotypes <- c("Focal motor seizure","HP:0000002","HP:0000003")
#### As phenotype names ####
pheno_names <- harmonise_phenotypes(phenotypes=phenotypes)
#### As HPO IDs ####
pheno_ids <- harmonise_phenotypes(phenotypes=phenotypes, as_hpo_ids=TRUE)
}
