% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_modifiers}
\alias{hpo_modifiers}
\title{Human Phenotype Ontology: disease modifiers}
\format{
data.table
}
\source{
\href{https://mseqdr.org/hpo_browser.php?12823;}{Severity order}

\href{https://hpo-annotation-qc.readthedocs.io/en/latest/annotationFormat.html}{Column explanations}

\code{
annot <- load_phenotype_to_genes(filename = "phenotype.hpoa")
annot <- annot[Modifier!=""]
parse_mod <- function(x){
unique(harmonise_phenotypes(strsplit(x,";")[[1]]))
}
annot <- annot[,Modifier_name:=lapply(Modifier,parse_mod)][Modifier!="",]
annot <- annot[,.(Modifier_name=unlist(Modifier_name)),
               by=c("HPO_ID","Modifier","DiseaseName","Aspect")]
dict <- c("Mild"=4,"Borderline"=3,"Severe"=2,"Profound"=1)
annot$Severity_score <- dict[annot$Modifier_name]
hpo_modifiers <- annot[,list(
  Modifier=paste(unique(Modifier),collapse=";"),
  Modifier_name=paste(unique(Modifier_name),collapse=";"),
  Modifier_count=paste(table(unlist(Modifier_name)),collapse=";"),
  DiseaseNames=paste(unique(DiseaseName),collapse = ";"),
  Severity_score_mean=mean(Severity_score, na.rm=TRUE),
  Severity_score_min=min(Severity_score, na.rm=TRUE)
), by="HPO_ID"]
hpo_modifiers[Severity_score_min<0,]$Severity_score_min <- NA
hpo_modifiers$Modifer_top <- lapply(seq_len(nrow(hpo_modifiers)),
                                    function(i){
  r <- hpo_modifiers[i,]
  mn <- strsplit(r$Modifier_name,";")[[1]]
  mc <- strsplit(r$Modifier_count,";")[[1]]
  mn[mc==max(mc)][[1]]
})
hpo_modifiers$Modifer_top <- unlist(hpo_modifiers$Modifer_top)
usethis::use_data(hpo_modifiers, overwrite = TRUE)
}
}
\usage{
data("hpo_modifiers")
}
\description{
Severity/progression modifiers of each HPO term.
In order of increasing severity:
\itemize{
\item{HP:0012825 }{Mild (Severity_score=4)}
\item{HP:0012827 }{Borderline (Severity_score=3)}
\item{HP:0012828 }{Severe (Severity_score=2)}
\item{HP:0012829 }{Profound (Severity_score=1)}
}
}
\keyword{datasets}
