% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_modifiers}
\alias{hpo_modifiers}
\title{Human Phenotype Ontology: disease modifiers}
\format{
data.table
}
\source{
\href{https://mseqdr.org/hpo_browser.php?12823;}{Severity order}

\href{https://hpo-annotation-qc.readthedocs.io/en/latest/annotationFormat.html}{Column explanations}

\code{
annot <- HPOExplorer::load_phenotype_to_genes(3)
annot <- annot[modifier!=""]
parse_mod <- function(x){
  unique(harmonise_phenotypes(strsplit(x,";")[[1]]))
}
annot <- annot[,modifier_name:=lapply(modifier,parse_mod)][modifier!="",]
annot <- annot[,.(modifier_name=unlist(modifier_name)),
               by=c("hpo_id","modifier","disease_name","aspect","disease_id")]
data.table::setnames(annot,"disease_id","disease_id")
dict <- HPOExplorer:::hpo_dict(type="severity")
annot$Severity_score <- dict[annot$modifier_name]
hpo_modifiers <- annot
usethis::use_data(hpo_modifiers, overwrite = TRUE)
}
}
\usage{
data("hpo_modifiers")
}
\description{
Severity/progression modifiers of each HPO term.
In order of increasing severity:
\itemize{
\item{HP:0012825 }{Mild (Severity_score=4)}
\item{HP:0012827 }{Borderline (Severity_score=3)}
\item{HP:0012828 }{Severe (Severity_score=2)}
\item{HP:0012829 }{Profound (Severity_score=1)}
}
}
\keyword{datasets}
