% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpo_onsets}
\alias{hpo_onsets}
\title{Human Phenotype Ontology: disease onset}
\format{
data.table
}
\source{
\code{
annot <- load_phenotype_to_genes(filename = "phenotype.hpoa")
annot <- annot[Onset!="",]
annot$Onset_name <- harmonise_phenotypes(phenotypes = annot$Onset,
                                         as_hpo_ids = FALSE)
counts <- dplyr::group_by(annot, DatabaseID) |>
  dplyr::summarise(HPO_IDs=length(unique(HPO_ID)),
                   Onsets=length(unique(Onset)))
## The number of Onsets partially depends on the number of HPO_IDs
## so it's necessary to keep HPO_ID too.
cor(counts$HPO_IDs, counts$Onsets)
hpo_onsets <- annot[,c("DatabaseID","HPO_ID","Onset","Onset_name")]
usethis::use_data(hpo_onsets, overwrite = TRUE)
}
}
\usage{
data("hpo_onsets")
}
\description{
HPO disease onsets.
}
\keyword{datasets}
