% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_phenotype_to_genes.R
\name{load_phenotype_to_genes}
\alias{load_phenotype_to_genes}
\title{Load phenotype_to_genes.txt}
\usage{
load_phenotype_to_genes(
  pheno_to_genes_txt_file = file.path(tools::R_user_dir("HPOExplorer", which = "cache"),
    "data", "phenotype_to_genes.txt")
)
}
\arguments{
\item{pheno_to_genes_txt_file}{path to the phenotype to genes text
file from the HPO. It contains phenotypes annotated with associated genes}
}
\value{
a dataframe of the phenotype_to_genes.txt file from HPO
}
\description{
This is a function for loading the phenotype to genes text file from HPO.
It adds the collumn names and returns as a dataframe
}
\examples{
phenotype_to_genes <- load_phenotype_to_genes()
}
