% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_phenotype_to_genes.R
\name{load_phenotype_to_genes}
\alias{load_phenotype_to_genes}
\title{Load phenotype_to_genes}
\source{
\href{https://hpo.jax.org/app/data/annotations}{HPO source data}
}
\usage{
load_phenotype_to_genes(
  filename = c("phenotype_to_genes.txt", "genes_to_phenotype.txt", "phenotype.hpoa"),
  save_dir = file.path(tools::R_user_dir("HPOExplorer", which = "cache"), "data"),
  method = "piggyback",
  verbose = TRUE
)
}
\arguments{
\item{filename}{Name of the file.}

\item{save_dir}{Folder where the phenotype to genes text
file is/will be stored.}

\item{method}{Where to download the data from:
\itemize{
\item{"piggyback": }{An old copy of HPO annotation files
stored in the Releases section of the \pkg{HPOExplorer} GitHub repository.}
\item{"hpo": }{An up-to-date copy directly from the official HPO website.}
}}

\item{verbose}{Print messages.}
}
\value{
A \link[data.table]{data.table} of the HPO annotations.
}
\description{
This is a function for loading the HPO phenotype to genes text file from HPO.
It adds the column names and returns as a dataframe.
It contains phenotypes annotated with associated genes.
}
\examples{
phenotype_to_genes <- load_phenotype_to_genes()
}
