% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_network_object.R
\name{make_network_object}
\alias{make_network_object}
\title{Make network object}
\usage{
make_network_object(
  phenos,
  hpo = get_hpo(),
  adjacency = adjacency_matrix(terms = phenos$HPO_ID, hpo = hpo),
  colour_var = "fold_change",
  add_ont_lvl_absolute = FALSE,
  cols = list_columns(extra_cols = c(colour_var, grep("_count$|_values$", names(phenos),
    value = TRUE))),
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{An adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{colour_var}{The column from phenos that you wish
to map to node colour.}

\item{add_ont_lvl_absolute}{Add the absolute ontology level of each HPO term.
See \link[HPOExplorer]{get_ont_lvls} for more details.}

\item{cols}{Columns to add to metadata of \link[ggnetwork]{ggnetwork} object.}

\item{verbose}{Print messages.}
}
\value{
A \link[ggnetwork]{ggnetwork} object
of a subset of the RD EWCE results.
}
\description{
This uses the network package to coerce the adjacency matrix into a
network object. It also adds the fold change, label,
and relative ontology level parameters to each node in the network.
}
\details{
It expects there to be a column of HPO IDs in the phenos dataframe called
HPO_ID.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenoNet <- make_network_object(phenos = phenos,
                                colour_var = "ontLvl_geneCount_ratio")
}
