% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_network_object.R
\name{make_network_object}
\alias{make_network_object}
\title{Make network object}
\usage{
make_network_object(
  phenos,
  hpo = get_hpo(),
  adjacency = adjacency_matrix(terms = phenos$HPO_ID, hpo = hpo),
  colour_var = "fold_change",
  add_ontLvl = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{adjacency}{AN adjacency matrix generated
by \link[HPOExplorer]{adjacency_matrix}.}

\item{colour_var}{The column from phenos that you wish
to map to node colour.}

\item{add_ontLvl}{Add the "ontLvl" (ontology level)
column if not already present.}

\item{verbose}{Print messages.}
}
\value{
A ggnetwork graph/ network object of a subset of the RD EWCE results.
}
\description{
This uses the network package to coerce the adjacency matrix into a
network object. It also adds the fold change, label,
and relative ontology level parameters to each node in the network.
}
\details{
It expects there to be a column of HPO IDs in the phenos dataframe called
HPO_ID.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
phenoNet <- make_network_object(phenos = phenos,
                                colour_var = "ontLvl_geneCount_ratio")
}
