% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_phenos_dataframe.R
\name{make_phenos_dataframe}
\alias{make_phenos_dataframe}
\title{Make phenotypes dataframe}
\usage{
make_phenos_dataframe(
  ancestor,
  hpo = get_hpo(),
  phenotype_to_genes = load_phenotype_to_genes(),
  add_description = TRUE,
  add_hoverboxes = TRUE,
  add_age_onset = FALSE,
  add_severity_tiers = FALSE,
  columns = list(Phenotype = "Phenotype", ID = "HPO_ID", ontLvl = "ontLvl", ontLvl_genes
    = "ontLvl_geneCount_ratio", Definition = "definition"),
  interactive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ancestor}{The ancestor to get all descendants of. If \code{NULL},
returns the entirely ontology.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{add_description}{Whether to get the phenotype descriptions as
well (slower).}

\item{add_hoverboxes}{Add hoverdata with
\link[HPOExplorer]{make_hoverboxes}.}

\item{add_age_onset}{Add age of onset columns using
\link[HPOExplorer]{add_onset}.}

\item{add_severity_tiers}{Add severity Tiers column using
\link[HPOExplorer]{add_tier}.}

\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[HPOExplorer]{make_hoverboxes}.}

\item{interactive}{Make the plot interactive with \link[plotly]{ggplotly}.}

\item{verbose}{Print messages.}
}
\value{
The HPO in dataframe format.
}
\description{
Make a dataframe from a subset of the Human Phenotype Ontology.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
}
