% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/list_deaths.R, R/list_onsets.R,
%   R/map_phenotypes.R
\name{map_}
\alias{map_}
\alias{list_deaths}
\alias{list_onsets}
\alias{map_phenotypes}
\title{Map functions}
\usage{
list_deaths(
  hpo = get_hpo(),
  exclude = FALSE,
  to = c("name", "id"),
  include_na = TRUE
)

list_onsets(
  hpo = get_hpo(),
  exclude = FALSE,
  to = c("name", "id"),
  include_na = TRUE,
  verbose = TRUE
)

map_phenotypes(
  terms,
  hpo = get_hpo(),
  to = c("name", "id"),
  keep_order = TRUE,
  invert = FALSE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{exclude}{HPO phenotype names to exclude.}

\item{to}{Character vector of database(s) to map IDs to.
When not \code{"mondo"}, can supply multiple alternative databases to map to
 (e.g. \code{c("OMIM","Orphanet","DECIPHER")}).}

\item{include_na}{Include NA values for onset.}

\item{verbose}{Print messages.}

\item{terms}{A subset of HPO IDs to include in the final dataset and plots
(e.g. c("HP:0001508","HP:0001507")).}

\item{keep_order}{Return a named list of the same length and order
as \code{terms}.
If \code{FALSE}, return a named list of only the unique \code{terms},
sometimes in a different order.}

\item{invert}{Invert the keys/values of the dictionary,
such that the key becomes the values (and vice versa).}
}
\value{
Mapped data.

Named list of HPO IDs.

Named list of HPO IDs.

Character vector
}
\description{
Functions to map IDs.
}
\section{Functions}{
\itemize{
\item \code{list_deaths()}: map_
List age of death HPO terms

List age of death phenotypes in the HPO.

\item \code{list_onsets()}: map_
List age of onset HPO terms

List age of onset phenotypes in the HPO.

\item \code{map_phenotypes()}: map_
Harmonise phenotypes

Harmonise a mixed vector of phenotype names (e.g. "Focal motor seizure")
and HPO IDs (e.g. c("HP:0000002","HP:0000003")).

}}
\examples{
deaths <- list_deaths()
onsets <- list_onsets()
terms <- c("Focal motor seizure","HP:0000002","HP:0000003")
#### As phenotype names ####
term_names <- map_phenotypes(terms=terms)
#### As HPO IDs ####
term_ids <- map_phenotypes(terms=terms, to="id")
}
\concept{map_}
