% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_3d.R
\name{network_3d}
\alias{network_3d}
\title{3D network}
\source{
\href{https://www.r-bloggers.com/2013/09/network-visualization-part-4-3d-networks/}{
R bloggers}

\href{https://community.plotly.com/t/is-it-possible-to-connect-scatters-in-3d-scatter-plot/}{
Plotly: Connect points in 3D plot}

\href{https://plotly.com/r/line-charts/}{
Plotly: Connect points in 2D plot}

\href{https://plotly.com/r/3d-line-plots/}{
Plotly: 3D lines plots}

\href{https://plotly.com/python/v3/3d-network-graph/}{
Plotly: 3D network plot}
}
\usage{
network_3d(
  g,
  layout_func = igraph::layout.fruchterman.reingold,
  node_color_var = "ontLvl",
  edge_color_var = "zend",
  text_color_var = node_color_var,
  node_symbol_var = "ancestor_name",
  node_palette = pals::ocean.phase,
  edge_palette = node_palette,
  kde_palette = pals::gnuplot,
  add_kde = TRUE,
  extend_kde = 1.5,
  bg_color = kde_palette(6)[1],
  add_labels = TRUE,
  keep_grid = FALSE,
  aspectmode = "cube",
  hover_width = 80,
  label_width = 80,
  seed = 2023,
  show_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{g}{\link[igraph]{igraph} object
generated by \link[HPOExplorer]{make_igraph_object}.}

\item{layout_func}{Layout function for the graph.}

\item{node_color_var}{Variable in the vertex metadata to color nodes by.}

\item{edge_color_var}{Variable in the edge metadata to color edges by.}

\item{text_color_var}{Variable in the node metadata to color text by.}

\item{node_symbol_var}{Variable in the vertex metadata to shape nodes by.}

\item{node_palette}{Color palette function for the nodes/points.}

\item{edge_palette}{Color palette function for the edges/lines.}

\item{kde_palette}{Color palette function for the KDE plot.}

\item{add_kde}{Add a kernel density estimation (KDE) plot
below the 3D scatter plot (i.e. the "mountains" beneath the points).}

\item{extend_kde}{Extend the area that the KDE plot covers.}

\item{bg_color}{Plot background color.}

\item{add_labels}{Add phenotype name labels to each point.}

\item{keep_grid}{Keep all grid lines and axis labels.}

\item{aspectmode}{The proportions of the 3D plot. See the
\href{https://plotly.com/python/reference/layout/scene/#layout-scene-aspectmode}{
plotly documentation site} for details.}

\item{hover_width}{Maximum width of the hover text.}

\item{label_width}{Maximum width of the label text.}

\item{seed}{Random seed to enable reproducibility.}

\item{show_plot}{Print the plot after it's been generated.}

\item{verbose}{Print messages.}
}
\value{
A 3D interactive \link[plotly]{plotly} object.
}
\description{
Plot a subset of the HPO as a 3D network.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
g <- make_igraph_object(phenos = phenos)
plt <- network_3d(g=g, show_plot=FALSE)
}
