% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenos_to_granges.R
\name{phenos_to_granges}
\alias{phenos_to_granges}
\title{Phenotypes to \link[GenomicRanges]{GenomicRanges}}
\usage{
phenos_to_granges(
  phenos,
  phenotype_to_genes = load_phenotype_to_genes(),
  hpo = get_hpo(),
  keep_seqnames = c(seq_len(22), "X", "Y"),
  split.field = "HPO_ID",
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{keep_seqnames}{Chromosomes to keep.}

\item{split.field}{
      A character string of a recognized column name in \code{df} that contains
      the grouping. This column defines how the rows of \code{df} are split and
      is typically a \code{factor} or \code{character} vector.  When
      \code{split.field} is not provided the \code{df} will be split by the
      number of rows.
  }

\item{verbose}{Print messages.}
}
\value{
A \link[GenomicRanges]{GRangesList}.
}
\description{
Convert a HPO phenotype dataframe generated by
\link[HPOExplorer]{make_phenos_dataframe}
to a \link[GenomicRanges]{GRangesList} split by HPO ID.
The resulting object will contain genes (and gene metadata) for all
genes associated with each phenotypes.
}
\examples{
phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
grl <- phenos_to_granges(phenos = phenos)
}
