% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_phenos.R
\name{plot_top_phenos}
\alias{plot_top_phenos}
\title{Plot top phenotypes}
\usage{
plot_top_phenos(
  res_class = gpt_annot_class(),
  keep_ont_levels = seq(3, 17),
  keep_descendants = "Phenotypic abnormality",
  n_per_class = 10,
  annotation_order = NULL,
  split_by_congenital = TRUE
)
}
\arguments{
\item{res_class}{Output of the \link{gpt_annot_class} function.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link{add_ont_lvl} for details.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{n_per_class}{Number of phenotypes per severity class to include.}

\item{annotation_order}{The order of the annotations to include.}

\item{split_by_congenital}{Split the phenotypes by congenital onset
(congenital = always/often, noncongenital = never/rarely).}
}
\description{
Plot the most severe phenotypes per severity class:
Profound, Severe, Moderate, Mild.
The exception to this is the "Mild" class, where the \emph{least}
severe phenotypes will be taken instead of the most severe phenotypes.
}
\examples{
res_class <- gpt_annot_class()
out <- plot_top_phenos(res_class = res_class)
}
