% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_hpo.R
\name{search_hpo}
\alias{search_hpo}
\title{Search the Human Phenotype Ontology}
\usage{
search_hpo(
  hpo = get_hpo(),
  queries = list(intellectual_disability = c("Intellectual disability",
    "Mental deterioration"), impaired_mobility = c("Gait disturbance",
    "Diminished movement", " mobility"), physical_malformations = c("malformation"),
    blindness = c("^blindness"), sensory_impairments = c("Abnormality of vision",
    "Abnormality of the sense of smell", "Abnormality of taste sensation",
    "Somatic sensory dysfunction", "Hearing abnormality"), immunodeficiency =
    c("Immunodeficiency", "Impaired antigen-specific response"), cancer = c("Cancer", 
  
      "malignant", "carcinoma"), reduced_fertility = c("Decreased fertility",
    "Hypogonadism")),
  search_cols = c("name"),
  include_descendants = TRUE,
  include_ancestors = FALSE,
  ignore.case = TRUE,
  fixed = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{queries}{A named list of character vectors.
Each vector includes a phenoptype name (or a substring)
to be searched for in the HPO.}

\item{search_cols}{HPO metadata column names to query for hits.}

\item{include_descendants}{Include all descendants of query hits.}

\item{include_ancestors}{Include all ancestors of query hits.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{fixed}{logical.  If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}

\item{verbose}{Print messages.}
}
\value{
Named list of HPO IDs.
}
\description{
Serarch for phenotypes in the HPO that match substring queries,
or are the descendants of those substring-matched phenotypes.
}
\examples{
query_hits <- search_hpo()
}
